@@ -1,58 +1,63 @@
 [metadata]
-url = https://github.com/poldracklab/niworkflows
-author = The CRN developers
-author_email = code@oscaresteban.es
-maintainer = Oscar Esteban
-maintainer_email = code@oscaresteban.es
-description = NeuroImaging Workflows provides processing tools for magnetic resonance images of the brain.
-long_description = file:README.rst
-long_description_content_type = text/x-rst; charset=UTF-8
-license = 3-clause BSD
+author = The NiPreps Developers
+author_email = nipreps@gmail.com
 classifiers =
     Development Status :: 4 - Beta
     Intended Audience :: Science/Research
     Topic :: Scientific/Engineering :: Image Recognition
     License :: OSI Approved :: BSD License
-    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+description = NeuroImaging Workflows provides processing tools for magnetic resonance images of the brain.
+license = 3-clause BSD
+license_file = LICENSE
+long_description = file:README.rst
+long_description_content_type = text/x-rst; charset=UTF-8
+project_urls =
+    Documentation = https://www.nipreps.org/niworkflows
+    GitHub = https://github.com/nipreps/niworkflows
+    Zenodo = https://doi.org/10.5281/zenodo.2650331
+url = https://github.com/nipreps/niworkflows
 
 [options]
-python_requires = >= 3.5
+python_requires = >= 3.6
 install_requires =
+    attrs
     jinja2
-    matplotlib >= 2.2.0 ; python_version >= "3.6"
-    matplotlib >= 2.2.0, < 3.1 ; python_version < "3.6"
-    nibabel ~= 3.0.0rc1
+    matplotlib >= 2.2.0
+    nibabel >= 3.0.1
     nilearn >= 0.2.6, != 0.5.0, != 0.5.1
     nipype >= 1.3.1
+    nitransforms >= 20.0.0rc3,<20.2
     packaging
     pandas
-    pybids >= 0.9.4
+    pybids >= 0.10.2
     PyYAML
-    scikit-image == 0.14.4 ; python_version < "3.6"
-    scikit-image ; python_version >= "3.6"
+    scikit-image
     scikit-learn
     scipy
     seaborn
     svgutils
     transforms3d
-    templateflow ~= 0.4.2rc1
+    templateflow >= 0.6
 test_requires =
+    coverage < 5
     pytest >= 4.4
-    pytest-xdist >= 1.28
     pytest-cov
-    coverage
+    pytest-xdist >= 1.28
 packages = find:
-include_package_data = True
+
+[options.entry_points]
+console_scripts =
+    niworkflows-boldref=niworkflows.cli.boldref:main
 
 [options.package_data]
 niworkflows =
     data/*.txt
-    data/t1-mni_registration*.json
-    data/bold-mni_registration*.json
-    reports/figures.json
-    reports/fmriprep.yml
+    data/*.json
+    data/sentinel.nii.gz
+    reports/default.yml
     reports/report.tpl
 
 [options.packages.find]
@@ -61,9 +66,9 @@ exclude =
 
 [options.extras_require]
 doc =
-    sphinx >= 2.1.2
     pydot >= 1.2.3
     pydotplus
+    sphinx >= 2.1.2, < 3.0
     sphinx_rtd_theme
     sphinxcontrib-apidoc ~= 0.3.0
     sphinxcontrib-napoleon
@@ -72,15 +77,16 @@ docs =
     %(doc)s
 duecredit =
     duecredit
+    citeproc-py != 0.5.0
 pointclouds =
     pyntcloud
 style =
     flake8 >= 3.7.0
 test =
+    coverage < 5
     pytest >= 4.4
-    pytest-xdist >= 1.28
     pytest-cov
-    coverage
+    pytest-xdist >= 1.28
 tests =
     %(test)s
 all =
@@ -101,6 +107,8 @@ parentdir_prefix =
 [flake8]
 max-line-length = 99
 doctests = False
+ignore =
+    W503
 exclude=*build/
 per-file-ignores =
     **/__init__.py : F401
@@ -109,7 +117,7 @@ per-file-ignores =
 [tool:pytest]
 norecursedirs = .git
 addopts = -svx --doctest-modules
-doctest_optionflags = ALLOW_UNICODE NORMALIZE_WHITESPACE
+doctest_optionflags = ALLOW_UNICODE NORMALIZE_WHITESPACE ELLIPSIS
 env =
     PYTHONHASHSEED=0
 filterwarnings =
