@@ -1,40 +1,74 @@
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
 """Interfaces for handling BIDS-like neuroimaging structures."""
-
-from json import dumps
+from collections import defaultdict
+from json import dumps, loads
 from pathlib import Path
 from shutil import copytree, rmtree
+from pkg_resources import resource_filename as _pkgres
+import re
 
 import nibabel as nb
+import numpy as np
+from bids.layout import parse_file_entities
+from bids.layout.writing import build_path
+from bids.utils import listify
 
 from nipype import logging
 from nipype.interfaces.base import (
-    traits, isdefined, Undefined,
-    TraitedSpec, BaseInterfaceInputSpec, DynamicTraitedSpec,
-    File, Directory, InputMultiObject, OutputMultiObject, Str,
+    traits,
+    isdefined,
+    Undefined,
+    TraitedSpec,
+    BaseInterfaceInputSpec,
+    DynamicTraitedSpec,
+    File,
+    Directory,
+    InputMultiObject,
+    OutputMultiObject,
+    Str,
     SimpleInterface,
 )
 from nipype.interfaces.io import add_traits
 from templateflow.api import templates as _get_template_list
-from ..utils.bids import BIDS_NAME, _init_layout
+from ..utils.bids import _init_layout, relative_to_root
 from ..utils.images import overwrite_header
 from ..utils.misc import splitext as _splitext, _copy_any
 
+regz = re.compile(r"\.gz$")
+_pybids_spec = loads(Path(_pkgres("niworkflows", "data/nipreps.json")).read_text())
+BIDS_DERIV_ENTITIES = frozenset({e["name"] for e in _pybids_spec["entities"]})
+BIDS_DERIV_PATTERNS = tuple(_pybids_spec["default_path_patterns"])
 
 STANDARD_SPACES = _get_template_list()
-LOGGER = logging.getLogger('nipype.interface')
+LOGGER = logging.getLogger("nipype.interface")
+
+
+def _none():
+    return None
+
+
+# Automatically coerce certain suffixes (DerivativesDataSink)
+DEFAULT_DTYPES = defaultdict(
+    _none,
+    (
+        ("mask", "uint8"),
+        ("dseg", "int16"),
+        ("probseg", "float32"),
+        ("boldref", "source"),
+    ),
+)
 
 
 class _BIDSBaseInputSpec(BaseInterfaceInputSpec):
     bids_dir = traits.Either(
-        (None, Directory(exists=True)), usedefault=True,
-        desc='optional bids directory to initialize a new layout')
-    bids_validate = traits.Bool(True, usedefault=True, desc='enable BIDS validator')
+        (None, Directory(exists=True)), usedefault=True, desc="optional bids directory"
+    )
+    bids_validate = traits.Bool(True, usedefault=True, desc="enable BIDS validator")
 
 
 class _BIDSInfoInputSpec(_BIDSBaseInputSpec):
-    in_file = File(mandatory=True, desc='input file, part of a BIDS tree')
+    in_file = File(mandatory=True, desc="input file, part of a BIDS tree")
 
 
 class _BIDSInfoOutputSpec(DynamicTraitedSpec):
@@ -130,37 +164,24 @@ sub-01/func/ses-retest/sub-01_ses-retest_task-covertverbgeneration_bold.nii.gz''
     task = covertverbgeneration
     <BLANKLINE>
 
-    >>> from bids import BIDSLayout
-    >>> bids_info = BIDSInfo()
-    >>> bids_info.layout = BIDSLayout(str(datadir / 'ds114'), validate=False)
-    >>> bids_info.inputs.in_file = '''\
-sub-01/func/ses-retest/sub-01_ses-retest_task-covertverbgeneration_bold.nii.gz'''
-    >>> res = bids_info.run()
-    >>> res.outputs
-    <BLANKLINE>
-    acquisition = <undefined>
-    reconstruction = <undefined>
-    run = <undefined>
-    session = retest
-    subject = 01
-    suffix = bold
-    task = covertverbgeneration
-    <BLANKLINE>
-
     """
 
     input_spec = _BIDSInfoInputSpec
     output_spec = _BIDSInfoOutputSpec
-    layout = None
 
     def _run_interface(self, runtime):
-        self.layout = self.inputs.bids_dir or self.layout
-        self.layout = _init_layout(self.inputs.in_file,
-                                   self.layout,
-                                   self.inputs.bids_validate)
-        params = self.layout.parse_file_entities(self.inputs.in_file)
-        self._results = {key: params.get(key, Undefined)
-                         for key in _BIDSInfoOutputSpec().get().keys()}
+        bids_dir = self.inputs.bids_dir
+        in_file = self.inputs.in_file
+        if bids_dir is not None:
+            try:
+                in_file = str(Path(in_file).relative_to(bids_dir))
+            except ValueError:
+                pass
+        params = parse_file_entities(in_file)
+        self._results = {
+            key: params.get(key, Undefined)
+            for key in _BIDSInfoOutputSpec().get().keys()
+        }
         return runtime
 
 
@@ -170,14 +191,14 @@ class _BIDSDataGrabberInputSpec(BaseInterfaceInputSpec):
 
 
 class _BIDSDataGrabberOutputSpec(TraitedSpec):
-    out_dict = traits.Dict(desc='output data structure')
-    fmap = OutputMultiObject(desc='output fieldmaps')
-    bold = OutputMultiObject(desc='output functional images')
-    sbref = OutputMultiObject(desc='output sbrefs')
-    t1w = OutputMultiObject(desc='output T1w images')
-    roi = OutputMultiObject(desc='output ROI images')
-    t2w = OutputMultiObject(desc='output T2w images')
-    flair = OutputMultiObject(desc='output FLAIR images')
+    out_dict = traits.Dict(desc="output data structure")
+    fmap = OutputMultiObject(desc="output fieldmaps")
+    bold = OutputMultiObject(desc="output functional images")
+    sbref = OutputMultiObject(desc="output sbrefs")
+    t1w = OutputMultiObject(desc="output T1w images")
+    roi = OutputMultiObject(desc="output ROI images")
+    t2w = OutputMultiObject(desc="output T2w images")
+    flair = OutputMultiObject(desc="output FLAIR images")
 
 
 class BIDSDataGrabber(SimpleInterface):
@@ -200,7 +221,7 @@ class BIDSDataGrabber(SimpleInterface):
     _require_funcs = True
 
     def __init__(self, *args, **kwargs):
-        anat_only = kwargs.pop('anat_only')
+        anat_only = kwargs.pop("anat_only")
         super(BIDSDataGrabber, self).__init__(*args, **kwargs)
         if anat_only is not None:
             self._require_funcs = not anat_only
@@ -208,82 +229,122 @@ class BIDSDataGrabber(SimpleInterface):
     def _run_interface(self, runtime):
         bids_dict = self.inputs.subject_data
 
-        self._results['out_dict'] = bids_dict
+        self._results["out_dict"] = bids_dict
         self._results.update(bids_dict)
 
-        if not bids_dict['t1w']:
-            raise FileNotFoundError('No T1w images found for subject sub-{}'.format(
-                self.inputs.subject_id))
+        if not bids_dict["t1w"]:
+            raise FileNotFoundError(
+                "No T1w images found for subject sub-{}".format(self.inputs.subject_id)
+            )
 
-        if self._require_funcs and not bids_dict['bold']:
-            raise FileNotFoundError('No functional images found for subject sub-{}'.format(
-                self.inputs.subject_id))
+        if self._require_funcs and not bids_dict["bold"]:
+            raise FileNotFoundError(
+                "No functional images found for subject sub-{}".format(
+                    self.inputs.subject_id
+                )
+            )
 
-        for imtype in ['bold', 't2w', 'flair', 'fmap', 'sbref', 'roi']:
+        for imtype in ["bold", "t2w", "flair", "fmap", "sbref", "roi"]:
             if not bids_dict[imtype]:
-                LOGGER.warning('No "%s" images found for sub-%s',
-                               imtype, self.inputs.subject_id)
+                LOGGER.info(
+                    'No "%s" images found for sub-%s', imtype, self.inputs.subject_id
+                )
 
         return runtime
 
 
 class _DerivativesDataSinkInputSpec(DynamicTraitedSpec, BaseInterfaceInputSpec):
     base_directory = traits.Directory(
-        desc='Path to the base directory for storing data.')
-    check_hdr = traits.Bool(True, usedefault=True, desc='fix headers of NIfTI outputs')
-    compress = traits.Bool(desc="force compression (True) or uncompression (False)"
-                                " of the output file (default: same as input)")
-    desc = Str('', usedefault=True, desc='Label for description field')
-    extra_values = traits.List(Str)
-    in_file = InputMultiObject(File(exists=True), mandatory=True,
-                               desc='the object to be saved')
-    keep_dtype = traits.Bool(False, usedefault=True, desc='keep datatype suffix')
-    meta_dict = traits.DictStrAny(desc='an input dictionary containing metadata')
-    source_file = File(exists=False, mandatory=True, desc='the input func file')
-    space = Str('', usedefault=True, desc='Label for space field')
-    suffix = Str('', usedefault=True, desc='suffix appended to source_file')
+        desc="Path to the base directory for storing data."
+    )
+    check_hdr = traits.Bool(True, usedefault=True, desc="fix headers of NIfTI outputs")
+    compress = InputMultiObject(
+        traits.Either(None, traits.Bool),
+        usedefault=True,
+        desc="whether ``in_file`` should be compressed (True), uncompressed (False) "
+        "or left unmodified (None, default).",
+    )
+    data_dtype = Str(
+        desc="NumPy datatype to coerce NIfTI data to, or `source` to"
+        "match the input file dtype"
+    )
+    dismiss_entities = InputMultiObject(
+        traits.Either(None, Str),
+        usedefault=True,
+        desc="a list entities that will not be propagated from the source file",
+    )
+    in_file = InputMultiObject(
+        File(exists=True), mandatory=True, desc="the object to be saved"
+    )
+    meta_dict = traits.DictStrAny(desc="an input dictionary containing metadata")
+    source_file = File(exists=False, mandatory=True, desc="the input func file")
 
 
 class _DerivativesDataSinkOutputSpec(TraitedSpec):
-    out_file = OutputMultiObject(File(exists=True, desc='written file path'))
-    out_meta = OutputMultiObject(File(exists=True, desc='written JSON sidecar path'))
+    out_file = OutputMultiObject(File(exists=True, desc="written file path"))
+    out_meta = OutputMultiObject(File(exists=True, desc="written JSON sidecar path"))
     compression = OutputMultiObject(
-        traits.Bool, desc='whether ``in_file`` was compressed/uncompressed '
-                          'or `it was copied directly.')
-    fixed_hdr = traits.List(traits.Bool, desc='whether derivative header was fixed')
+        traits.Either(None, traits.Bool),
+        desc="whether ``in_file`` should be compressed (True), uncompressed (False) "
+        "or left unmodified (None).",
+    )
+    fixed_hdr = traits.List(traits.Bool, desc="whether derivative header was fixed")
 
 
 class DerivativesDataSink(SimpleInterface):
     """
     Store derivative files.
 
-    Saves the `in_file` into a BIDS-Derivatives folder provided
-    by `base_directory`, given the input reference `source_file`.
+    Saves the ``in_file`` into a BIDS-Derivatives folder provided
+    by ``base_directory``, given the input reference ``source_file``.
 
     >>> import tempfile
     >>> tmpdir = Path(tempfile.mkdtemp())
     >>> tmpfile = tmpdir / 'a_temp_file.nii.gz'
     >>> tmpfile.open('w').close()  # "touch" the file
+    >>> t1w_source = bids_collect_data(
+    ...     str(datadir / 'ds114'), '01', bids_validate=False)[0]['t1w'][0]
     >>> dsink = DerivativesDataSink(base_directory=str(tmpdir), check_hdr=False)
     >>> dsink.inputs.in_file = str(tmpfile)
-    >>> dsink.inputs.source_file = bids_collect_data(
-    ...     str(datadir / 'ds114'), '01', bids_validate=False)[0]['t1w'][0]
-    >>> dsink.inputs.keep_dtype = True
-    >>> dsink.inputs.suffix = 'desc-denoised'
+    >>> dsink.inputs.source_file = t1w_source
+    >>> dsink.inputs.desc = 'denoised'
+    >>> dsink.inputs.compress = False
     >>> res = dsink.run()
     >>> res.outputs.out_file  # doctest: +ELLIPSIS
-    '.../niworkflows/sub-01/ses-retest/anat/sub-01_ses-retest_desc-denoised_T1w.nii.gz'
+    '.../niworkflows/sub-01/ses-retest/anat/sub-01_ses-retest_desc-denoised_T1w.nii'
+
+    >>> tmpfile = tmpdir / 'a_temp_file.nii'
+    >>> tmpfile.open('w').close()  # "touch" the file
+    >>> dsink = DerivativesDataSink(base_directory=str(tmpdir), check_hdr=False,
+    ...                             allowed_entities=("custom",))
+    >>> dsink.inputs.in_file = str(tmpfile)
+    >>> dsink.inputs.source_file = t1w_source
+    >>> dsink.inputs.custom = 'noise'
+    >>> res = dsink.run()
+    >>> res.outputs.out_file  # doctest: +ELLIPSIS
+    '.../niworkflows/sub-01/ses-retest/anat/sub-01_ses-retest_custom-noise_T1w.nii'
 
     >>> dsink = DerivativesDataSink(base_directory=str(tmpdir), check_hdr=False,
-    ...                             allowed_entities=['from', 'to'], **{'from': 'orig'})
-    >>> dsink.inputs.in_file = str(tmpfile)
-    >>> dsink.inputs.to = 'native'
-    >>> dsink.inputs.source_file = bids_collect_data(
-    ...     str(datadir / 'ds114'), '01', bids_validate=False)[0]['t1w'][0]
-    >>> dsink.inputs.keep_dtype = True
+    ...                             allowed_entities=("custom",))
+    >>> dsink.inputs.in_file = [str(tmpfile), str(tmpfile)]
+    >>> dsink.inputs.source_file = t1w_source
+    >>> dsink.inputs.custom = [1, 2]
+    >>> dsink.inputs.compress = True
     >>> res = dsink.run()
     >>> res.outputs.out_file  # doctest: +ELLIPSIS
-    '.../sub-01_ses-retest_from-orig_to-native_T1w.nii.gz'
+    ['.../niworkflows/sub-01/ses-retest/anat/sub-01_ses-retest_custom-1_T1w.nii.gz',
+     '.../niworkflows/sub-01/ses-retest/anat/sub-01_ses-retest_custom-2_T1w.nii.gz']
+
+    >>> dsink = DerivativesDataSink(base_directory=str(tmpdir), check_hdr=False,
+    ...                             allowed_entities=("custom1", "custom2"))
+    >>> dsink.inputs.in_file = [str(tmpfile)] * 2
+    >>> dsink.inputs.source_file = t1w_source
+    >>> dsink.inputs.custom1 = [1, 2]
+    >>> dsink.inputs.custom2 = "b"
+    >>> res = dsink.run()
+    >>> res.outputs.out_file  # doctest: +ELLIPSIS
+    ['.../niworkflows/sub-01/ses-retest/anat/sub-01_ses-retest_custom1-1_custom2-b_T1w.nii',
+     '.../niworkflows/sub-01/ses-retest/anat/sub-01_ses-retest_custom1-2_custom2-b_T1w.nii']
 
     >>> bids_dir = tmpdir / 'bidsroot' / 'sub-02' / 'ses-noanat' / 'func'
     >>> bids_dir.mkdir(parents=True, exist_ok=True)
@@ -292,26 +353,24 @@ class DerivativesDataSink(SimpleInterface):
     >>> dsink = DerivativesDataSink(base_directory=str(tmpdir), check_hdr=False)
     >>> dsink.inputs.in_file = str(tmpfile)
     >>> dsink.inputs.source_file = str(tricky_source)
-    >>> dsink.inputs.keep_dtype = True
     >>> dsink.inputs.desc = 'preproc'
     >>> res = dsink.run()
     >>> res.outputs.out_file  # doctest: +ELLIPSIS
-    '.../niworkflows/sub-02/ses-noanat/func/sub-02_ses-noanat_task-rest_run-01_\
-desc-preproc_bold.nii.gz'
+    '.../niworkflows/sub-02/ses-noanat/func/sub-02_ses-noanat_task-rest_run-1_\
+desc-preproc_bold.nii'
 
     >>> bids_dir = tmpdir / 'bidsroot' / 'sub-02' / 'ses-noanat' / 'func'
     >>> bids_dir.mkdir(parents=True, exist_ok=True)
-    >>> tricky_source = bids_dir / 'sub-02_ses-noanat_task-rest_run-01_bold.nii.gz'
+    >>> tricky_source = bids_dir / 'sub-02_ses-noanat_task-rest_run-1_bold.nii.gz'
     >>> tricky_source.open('w').close()
     >>> dsink = DerivativesDataSink(base_directory=str(tmpdir), check_hdr=False)
     >>> dsink.inputs.in_file = str(tmpfile)
     >>> dsink.inputs.source_file = str(tricky_source)
-    >>> dsink.inputs.keep_dtype = True
     >>> dsink.inputs.desc = 'preproc'
     >>> dsink.inputs.RepetitionTime = 0.75
     >>> res = dsink.run()
     >>> res.outputs.out_meta  # doctest: +ELLIPSIS
-    '.../niworkflows/sub-02/ses-noanat/func/sub-02_ses-noanat_task-rest_run-01_\
+    '.../niworkflows/sub-02/ses-noanat/func/sub-02_ses-noanat_task-rest_run-1_\
 desc-preproc_bold.json'
 
     >>> Path(res.outputs.out_meta).read_text().splitlines()[1]
@@ -325,13 +384,14 @@ desc-preproc_bold.json'
     ...                             SkullStripped=True)
     >>> dsink.inputs.in_file = str(tmpfile)
     >>> dsink.inputs.source_file = str(tricky_source)
-    >>> dsink.inputs.keep_dtype = True
     >>> dsink.inputs.desc = 'preproc'
+    >>> dsink.inputs.space = 'MNI152NLin6Asym'
+    >>> dsink.inputs.resolution = '01'
     >>> dsink.inputs.RepetitionTime = 0.75
     >>> res = dsink.run()
     >>> res.outputs.out_meta  # doctest: +ELLIPSIS
-    '.../niworkflows/sub-02/ses-noanat/func/sub-02_ses-noanat_task-rest_run-01_\
-desc-preproc_bold.json'
+    '.../niworkflows/sub-02/ses-noanat/func/sub-02_ses-noanat_task-rest_run-1_\
+space-MNI152NLin6Asym_res-01_desc-preproc_bold.json'
 
     >>> lines = Path(res.outputs.out_meta).read_text().splitlines()
     >>> lines[1]
@@ -348,14 +408,15 @@ desc-preproc_bold.json'
     ...                             SkullStripped=True)
     >>> dsink.inputs.in_file = str(tmpfile)
     >>> dsink.inputs.source_file = str(tricky_source)
-    >>> dsink.inputs.keep_dtype = True
     >>> dsink.inputs.desc = 'preproc'
+    >>> dsink.inputs.resolution = 'native'
+    >>> dsink.inputs.space = 'MNI152NLin6Asym'
     >>> dsink.inputs.RepetitionTime = 0.75
     >>> dsink.inputs.meta_dict = {'RepetitionTime': 1.75, 'SkullStripped': False, 'Z': 'val'}
     >>> res = dsink.run()
     >>> res.outputs.out_meta  # doctest: +ELLIPSIS
-    '.../niworkflows/sub-02/ses-noanat/func/sub-02_ses-noanat_task-rest_run-01_\
-desc-preproc_bold.json'
+    '.../niworkflows/sub-02/ses-noanat/func/sub-02_ses-noanat_task-rest_run-1_\
+space-MNI152NLin6Asym_desc-preproc_bold.json'
 
     >>> lines = Path(res.outputs.out_meta).read_text().splitlines()
     >>> lines[1]
@@ -373,141 +434,203 @@ desc-preproc_bold.json'
     output_spec = _DerivativesDataSinkOutputSpec
     out_path_base = "niworkflows"
     _always_run = True
+    _allowed_entities = set(BIDS_DERIV_ENTITIES)
 
     def __init__(self, allowed_entities=None, out_path_base=None, **inputs):
-        self._allowed_entities = allowed_entities or []
-
-        self._metadata = {}
-        self._static_traits = self.input_spec.class_editable_traits() + self._allowed_entities
-        for dynamic_input in set(inputs) - set(self._static_traits):
-            self._metadata[dynamic_input] = inputs.pop(dynamic_input)
-
-        super(DerivativesDataSink, self).__init__(**inputs)
-        if self._allowed_entities:
-            add_traits(self.inputs, self._allowed_entities)
-            for k in set(self._allowed_entities).intersection(list(inputs.keys())):
-                setattr(self.inputs, k, inputs[k])
-
-        self._results['out_file'] = []
+        """Initialize the SimpleInterface and extend inputs with custom entities."""
+        self._allowed_entities = set(allowed_entities or []).union(
+            self._allowed_entities
+        )
         if out_path_base:
             self.out_path_base = out_path_base
 
+        self._metadata = {}
+        self._static_traits = self.input_spec.class_editable_traits() + sorted(
+            self._allowed_entities
+        )
+        for dynamic_input in set(inputs) - set(self._static_traits):
+            self._metadata[dynamic_input] = inputs.pop(dynamic_input)
+
+        # First regular initialization (constructs InputSpec object)
+        super().__init__(**inputs)
+        add_traits(self.inputs, self._allowed_entities)
+        for k in self._allowed_entities.intersection(list(inputs.keys())):
+            # Add additional input fields (self.inputs is an object)
+            setattr(self.inputs, k, inputs[k])
+
     def _run_interface(self, runtime):
+        # Ready the output folder
+        base_directory = runtime.cwd
+        if isdefined(self.inputs.base_directory):
+            base_directory = self.inputs.base_directory
+        base_directory = Path(base_directory).absolute()
+        out_path = base_directory / self.out_path_base
+        out_path.mkdir(exist_ok=True, parents=True)
+
+        # Ensure we have a list
+        in_file = listify(self.inputs.in_file)
+
+        # Read in the dictionary of metadata
         if isdefined(self.inputs.meta_dict):
             meta = self.inputs.meta_dict
             # inputs passed in construction take priority
             meta.update(self._metadata)
             self._metadata = meta
 
-        src_fname, _ = _splitext(self.inputs.source_file)
-        src_fname, dtype = src_fname.rsplit('_', 1)
-        _, ext = _splitext(self.inputs.in_file[0])
-        if self.inputs.compress is True and not ext.endswith('.gz'):
-            ext += '.gz'
-        elif self.inputs.compress is False and ext.endswith('.gz'):
-            ext = ext[:-3]
+        # Initialize entities with those from the source file.
+        out_entities = parse_file_entities(
+            str(relative_to_root(self.inputs.source_file))
+        )
+        for drop_entity in listify(self.inputs.dismiss_entities or []):
+            out_entities.pop(drop_entity, None)
 
-        m = BIDS_NAME.search(src_fname)
+        # Override extension with that of the input file(s)
+        out_entities["extension"] = [
+            # _splitext does not accept .surf.gii (for instance)
+            "".join(Path(orig_file).suffixes).lstrip(".")
+            for orig_file in in_file
+        ]
 
-        mod = Path(self.inputs.source_file).parent.name
+        compress = listify(self.inputs.compress) or [None]
+        if len(compress) == 1:
+            compress = compress * len(in_file)
+        for i, ext in enumerate(out_entities["extension"]):
+            if compress[i] is not None:
+                ext = regz.sub("", ext)
+                out_entities["extension"][i] = f"{ext}.gz" if compress[i] else ext
 
-        base_directory = runtime.cwd
-        if isdefined(self.inputs.base_directory):
-            base_directory = self.inputs.base_directory
-
-        base_directory = Path(base_directory).resolve()
-        out_path = base_directory / self.out_path_base / \
-            '{subject_id}'.format(**m.groupdict())
-
-        if m.groupdict().get('session_id') is not None:
-            out_path = out_path / '{session_id}'.format(**m.groupdict())
-
-        out_path = out_path / '{}'.format(mod)
-        out_path.mkdir(exist_ok=True, parents=True)
-        base_fname = str(out_path / src_fname)
-
-        allowed_entities = {}
+        # Override entities with those set as inputs
         for key in self._allowed_entities:
             value = getattr(self.inputs, key)
             if value is not None and isdefined(value):
-                allowed_entities[key] = '_%s-%s' % (key, value)
+                out_entities[key] = value
 
-        formatbase = '{bname}{space}{desc}' + ''.join(
-            [allowed_entities.get(s, '') for s in self._allowed_entities])
+        # Clean up native resolution with space
+        if out_entities.get("resolution") == "native" and out_entities.get("space"):
+            out_entities.pop("resolution", None)
 
-        formatstr = formatbase + '{extra}{suffix}{dtype}{ext}'
-        if len(self.inputs.in_file) > 1 and not isdefined(self.inputs.extra_values):
-            formatstr = formatbase + '{suffix}{i:04d}{dtype}{ext}'
+        if len(set(out_entities["extension"])) == 1:
+            out_entities["extension"] = out_entities["extension"][0]
 
-        space = '_space-{}'.format(self.inputs.space) if self.inputs.space else ''
-        desc = '_desc-{}'.format(self.inputs.desc) if self.inputs.desc else ''
-        suffix = '_{}'.format(self.inputs.suffix) if self.inputs.suffix else ''
-        dtype = '' if not self.inputs.keep_dtype else ('_%s' % dtype)
+        # Insert custom (non-BIDS) entities from allowed_entities.
+        custom_entities = set(out_entities.keys()) - set(BIDS_DERIV_ENTITIES)
+        patterns = BIDS_DERIV_PATTERNS
+        if custom_entities:
+            # Example: f"{key}-{{{key}}}" -> "task-{task}"
+            custom_pat = "_".join(f"{key}-{{{key}}}" for key in sorted(custom_entities))
+            patterns = [
+                pat.replace("_{suffix", "_".join(("", custom_pat, "{suffix")))
+                for pat in patterns
+            ]
 
-        self._results['compression'] = []
-        self._results['fixed_hdr'] = [False] * len(self.inputs.in_file)
+        # Prepare SimpleInterface outputs object
+        self._results["out_file"] = []
+        self._results["compression"] = []
+        self._results["fixed_hdr"] = [False] * len(in_file)
 
-        for i, fname in enumerate(self.inputs.in_file):
-            extra = ''
-            if isdefined(self.inputs.extra_values):
-                extra = '_{}'.format(self.inputs.extra_values[i])
-            out_file = formatstr.format(
-                bname=base_fname,
-                space=space,
-                desc=desc,
-                extra=extra,
-                suffix=suffix,
-                i=i,
-                dtype=dtype,
-                ext=ext,
+        dest_files = build_path(out_entities, path_patterns=patterns)
+        if not dest_files:
+            raise ValueError(f"Could not build path with entities {out_entities}.")
+
+        # Make sure the interpolated values is embedded in a list, and check
+        dest_files = listify(dest_files)
+        if len(in_file) != len(dest_files):
+            raise ValueError(
+                f"Input files ({len(in_file)}) not matched "
+                f"by interpolated patterns ({len(dest_files)})."
             )
-            self._results['out_file'].append(out_file)
-            self._results['compression'].append(_copy_any(fname, out_file))
 
-            is_nii = out_file.endswith('.nii') or out_file.endswith('.nii.gz')
-            if self.inputs.check_hdr and is_nii:
+        for i, (orig_file, dest_file) in enumerate(zip(in_file, dest_files)):
+            out_file = out_path / dest_file
+            out_file.parent.mkdir(exist_ok=True, parents=True)
+            self._results["out_file"].append(str(out_file))
+            self._results["compression"].append(_copy_any(orig_file, str(out_file)))
+
+            is_nifti = out_file.name.endswith(
+                (".nii", ".nii.gz")
+            ) and not out_file.name.endswith((".dtseries.nii", ".dtseries.nii.gz"))
+            data_dtype = self.inputs.data_dtype or DEFAULT_DTYPES[self.inputs.suffix]
+            if is_nifti and any((self.inputs.check_hdr, data_dtype)):
                 # Do not use mmap; if we need to access the data at all, it will be to
                 # rewrite, risking a BusError
                 nii = nb.load(out_file, mmap=False)
-                if not isinstance(nii, (nb.Nifti1Image, nb.Nifti2Image)):
-                    # .dtseries.nii are CIfTI2, therefore skip check
-                    return runtime
-                hdr = nii.header
-                curr_units = tuple([None if u == 'unknown' else u
-                                    for u in hdr.get_xyzt_units()])
-                curr_codes = (int(hdr['qform_code']), int(hdr['sform_code']))
 
-                # Default to mm, use sec if data type is bold
-                units = (curr_units[0] or 'mm', 'sec' if dtype == '_bold' else None)
-                xcodes = (1, 1)  # Derivative in its original scanner space
-                if self.inputs.space:
-                    xcodes = (4, 4) if self.inputs.space in STANDARD_SPACES \
-                        else (2, 2)
+                if self.inputs.check_hdr:
+                    hdr = nii.header
+                    curr_units = tuple(
+                        [None if u == "unknown" else u for u in hdr.get_xyzt_units()]
+                    )
+                    curr_codes = (int(hdr["qform_code"]), int(hdr["sform_code"]))
 
-                if curr_codes != xcodes or curr_units != units:
-                    self._results['fixed_hdr'][i] = True
-                    hdr.set_qform(nii.affine, xcodes[0])
-                    hdr.set_sform(nii.affine, xcodes[1])
-                    hdr.set_xyzt_units(*units)
+                    # Default to mm, use sec if data type is bold
+                    units = (
+                        curr_units[0] or "mm",
+                        "sec" if out_entities["suffix"] == "bold" else None,
+                    )
+                    xcodes = (1, 1)  # Derivative in its original scanner space
+                    if self.inputs.space:
+                        xcodes = (
+                            (4, 4) if self.inputs.space in STANDARD_SPACES else (2, 2)
+                        )
 
-                    # Rewrite file with new header
-                    overwrite_header(nii, out_file)
+                    if curr_codes != xcodes or curr_units != units:
+                        self._results["fixed_hdr"][i] = True
+                        hdr.set_qform(nii.affine, xcodes[0])
+                        hdr.set_sform(nii.affine, xcodes[1])
+                        hdr.set_xyzt_units(*units)
 
-        if len(self._results['out_file']) == 1:
+                        # Rewrite file with new header
+                        overwrite_header(nii, out_file)
+
+                if data_dtype == "source":  # match source dtype
+                    try:
+                        data_dtype = nb.load(self.inputs.source_file).get_data_dtype()
+                    except Exception:
+                        LOGGER.warning(
+                            f"Could not get data type of file {self.inputs.source_file}"
+                        )
+                        data_dtype = None
+
+                if data_dtype:
+                    if self.inputs.check_hdr:
+                        # load updated NIfTI
+                        nii = nb.load(out_file, mmap=False)
+                    data_dtype = np.dtype(data_dtype)
+                    orig_dtype = nii.get_data_dtype()
+                    if orig_dtype != data_dtype:
+                        LOGGER.warning(
+                            f"Changing {out_file} dtype from {orig_dtype} to {data_dtype}"
+                        )
+                        # coerce dataobj to new data dtype
+                        if np.issubdtype(data_dtype, np.integer):
+                            new_data = np.rint(nii.dataobj).astype(data_dtype)
+                        else:
+                            new_data = np.asanyarray(nii.dataobj, dtype=data_dtype)
+                        # and set header to match
+                        nii.set_data_dtype(data_dtype)
+                        nii = nii.__class__(new_data, nii.affine, nii.header)
+                        nii.to_filename(out_file)
+
+        if len(self._results["out_file"]) == 1:
             meta_fields = self.inputs.copyable_trait_names()
-            self._metadata.update({
-                k: getattr(self.inputs, k)
-                for k in meta_fields if k not in self._static_traits})
+            self._metadata.update(
+                {
+                    k: getattr(self.inputs, k)
+                    for k in meta_fields
+                    if k not in self._static_traits
+                }
+            )
             if self._metadata:
-                sidecar = (Path(self._results['out_file'][0]).parent /
-                           ('%s.json' % _splitext(self._results['out_file'][0])[0]))
+                sidecar = Path(self._results["out_file"][0]).parent / (
+                    "%s.json" % _splitext(self._results["out_file"][0])[0]
+                )
                 sidecar.write_text(dumps(self._metadata, sort_keys=True, indent=2))
-                self._results['out_meta'] = str(sidecar)
+                self._results["out_meta"] = str(sidecar)
         return runtime
 
 
 class _ReadSidecarJSONInputSpec(_BIDSBaseInputSpec):
-    in_file = File(exists=True, mandatory=True, desc='the input nifti file')
+    in_file = File(exists=True, mandatory=True, desc="the input nifti file")
 
 
 class _ReadSidecarJSONOutputSpec(_BIDSInfoOutputSpec):
@@ -561,10 +684,7 @@ class ReadSidecarJSON(SimpleInterface):
 
     def __init__(self, fields=None, undef_fields=False, **inputs):
         super(ReadSidecarJSON, self).__init__(**inputs)
-        self._fields = fields or []
-        if isinstance(self._fields, str):
-            self._fields = [self._fields]
-
+        self._fields = listify(fields or [])
         self._undef_fields = undef_fields
 
     def _outputs(self):
@@ -575,48 +695,54 @@ class ReadSidecarJSON(SimpleInterface):
 
     def _run_interface(self, runtime):
         self.layout = self.inputs.bids_dir or self.layout
-        self.layout = _init_layout(self.inputs.in_file,
-                                   self.layout,
-                                   self.inputs.bids_validate)
+        self.layout = _init_layout(
+            self.inputs.in_file, self.layout, self.inputs.bids_validate
+        )
 
         # Fill in BIDS entities of the output ("*_id")
         output_keys = list(_BIDSInfoOutputSpec().get().keys())
         params = self.layout.parse_file_entities(self.inputs.in_file)
-        self._results = {key: params.get(key.split('_')[0], Undefined)
-                         for key in output_keys}
+        self._results = {
+            key: params.get(key.split("_")[0], Undefined) for key in output_keys
+        }
 
         # Fill in metadata
         metadata = self.layout.get_metadata(self.inputs.in_file)
-        self._results['out_dict'] = metadata
+        self._results["out_dict"] = metadata
 
         # Set dynamic outputs if fields input is present
         for fname in self._fields:
             if not self._undef_fields and fname not in metadata:
                 raise KeyError(
-                    'Metadata field "%s" not found for file %s' % (
-                        fname, self.inputs.in_file))
+                    'Metadata field "%s" not found for file %s'
+                    % (fname, self.inputs.in_file)
+                )
             self._results[fname] = metadata.get(fname, Undefined)
         return runtime
 
 
 class _BIDSFreeSurferDirInputSpec(BaseInterfaceInputSpec):
-    derivatives = Directory(exists=True, mandatory=True,
-                            desc='BIDS derivatives directory')
-    freesurfer_home = Directory(exists=True, mandatory=True,
-                                desc='FreeSurfer installation directory')
-    subjects_dir = traits.Either(traits.Str(),
-                                 Directory(),
-                                 default='freesurfer',
-                                 usedefault=True,
-                                 desc='Name of FreeSurfer subjects directory')
-    spaces = traits.List(traits.Str, desc='Set of output spaces to prepare')
-    overwrite_fsaverage = traits.Bool(False, usedefault=True,
-                                      desc='Overwrite fsaverage directories, if present')
+    derivatives = Directory(
+        exists=True, mandatory=True, desc="BIDS derivatives directory"
+    )
+    freesurfer_home = Directory(
+        exists=True, mandatory=True, desc="FreeSurfer installation directory"
+    )
+    subjects_dir = traits.Either(
+        traits.Str(),
+        Directory(),
+        default="freesurfer",
+        usedefault=True,
+        desc="Name of FreeSurfer subjects directory",
+    )
+    spaces = traits.List(traits.Str, desc="Set of output spaces to prepare")
+    overwrite_fsaverage = traits.Bool(
+        False, usedefault=True, desc="Overwrite fsaverage directories, if present"
+    )
 
 
 class _BIDSFreeSurferDirOutputSpec(TraitedSpec):
-    subjects_dir = traits.Directory(exists=True,
-                                    desc='FreeSurfer subjects directory')
+    subjects_dir = traits.Directory(exists=True, desc="FreeSurfer subjects directory")
 
 
 class BIDSFreeSurferDir(SimpleInterface):
@@ -649,9 +775,9 @@ class BIDSFreeSurferDir(SimpleInterface):
         if not subjects_dir.is_absolute():
             subjects_dir = Path(self.inputs.derivatives) / subjects_dir
         subjects_dir.mkdir(parents=True, exist_ok=True)
-        self._results['subjects_dir'] = str(subjects_dir)
+        self._results["subjects_dir"] = str(subjects_dir)
 
-        orig_subjects_dir = Path(self.inputs.freesurfer_home) / 'subjects'
+        orig_subjects_dir = Path(self.inputs.freesurfer_home) / "subjects"
 
         # Source is target, so just quit
         if subjects_dir == orig_subjects_dir:
@@ -659,12 +785,12 @@ class BIDSFreeSurferDir(SimpleInterface):
 
         spaces = list(self.inputs.spaces)
         # Always copy fsaverage, for proper recon-all functionality
-        if 'fsaverage' not in spaces:
-            spaces.append('fsaverage')
+        if "fsaverage" not in spaces:
+            spaces.append("fsaverage")
 
         for space in spaces:
             # Skip non-freesurfer spaces and fsnative
-            if not space.startswith('fsaverage'):
+            if not space.startswith("fsaverage"):
                 continue
             source = orig_subjects_dir / space
             dest = subjects_dir / space
@@ -683,7 +809,10 @@ class BIDSFreeSurferDir(SimpleInterface):
                 try:
                     copytree(source, dest)
                 except FileExistsError:
-                    LOGGER.warning("%s exists; if multiple jobs are running in parallel"
-                                   ", this can be safely ignored", dest)
+                    LOGGER.warning(
+                        "%s exists; if multiple jobs are running in parallel"
+                        ", this can be safely ignored",
+                        dest,
+                    )
 
         return runtime
