@@ -23,7 +23,7 @@ sys.path.append(os.path.abspath('sphinxext'))
 # -- Project information -----------------------------------------------------
 project = __packagename__
 copyright = __copyright__
-author = 'The Niworkflows Developers'
+author = 'The NiPreps Developers'
 
 # The short X.Y version
 version = Version(__version__).base_version
@@ -156,8 +156,8 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'smriprep.tex', 'sMRIPrep Documentation',
-     'The sMRIPrep Developers', 'manual'),
+    (master_doc, 'niworkflows.tex', 'NiWorkflows Documentation',
+     'The NiPreps Developers', 'manual'),
 ]
 
 
@@ -166,7 +166,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'smriprep', 'sMRIPrep Documentation',
+    (master_doc, 'niworkflows', 'NiWorkflows Documentation',
      [author], 1)
 ]
 
@@ -177,8 +177,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'smriprep', 'sMRIPrep Documentation',
-     author, 'sMRIPrep', 'One line description of project.',
+    (master_doc, 'niworkflows', 'NiWorkflows Documentation',
+     author, 'NiWorkflows', 'One line description of project.',
      'Miscellaneous'),
 ]
 
