@@ -7,7 +7,7 @@ from ..interfaces.nibabel import Binarize
 from ..interfaces.fixes import FixN4BiasFieldCorrection as N4BiasFieldCorrection
 
 
-def afni_wf(name='AFNISkullStripWorkflow', unifize=False, n4_nthreads=1):
+def afni_wf(name="AFNISkullStripWorkflow", unifize=False, n4_nthreads=1):
     """
     Create a skull-stripping workflow based on AFNI's tools.
 
@@ -52,50 +52,68 @@ def afni_wf(name='AFNISkullStripWorkflow', unifize=False, n4_nthreads=1):
 
     """
     workflow = pe.Workflow(name=name)
-    inputnode = pe.Node(niu.IdentityInterface(fields=['in_file']),
-                        name='inputnode')
-    outputnode = pe.Node(niu.IdentityInterface(
-        fields=['bias_corrected', 'out_file', 'out_mask', 'bias_image']), name='outputnode')
+    inputnode = pe.Node(niu.IdentityInterface(fields=["in_file"]), name="inputnode")
+    outputnode = pe.Node(
+        niu.IdentityInterface(
+            fields=["bias_corrected", "out_file", "out_mask", "bias_image"]
+        ),
+        name="outputnode",
+    )
 
     inu_n4 = pe.Node(
-        N4BiasFieldCorrection(dimension=3, save_bias=True, num_threads=n4_nthreads,
-                              rescale_intensities=True, copy_header=True),
+        N4BiasFieldCorrection(
+            dimension=3,
+            save_bias=True,
+            num_threads=n4_nthreads,
+            rescale_intensities=True,
+            copy_header=True,
+        ),
         n_procs=n4_nthreads,
-        name='inu_n4')
+        name="inu_n4",
+    )
 
-    sstrip = pe.Node(afni.SkullStrip(outputtype='NIFTI_GZ'), name='skullstrip')
-    sstrip_orig_vol = pe.Node(afni.Calc(
-        expr='a*step(b)', outputtype='NIFTI_GZ'), name='sstrip_orig_vol')
-    binarize = pe.Node(Binarize(thresh_low=0.0), name='binarize')
+    sstrip = pe.Node(afni.SkullStrip(outputtype="NIFTI_GZ"), name="skullstrip")
+    sstrip_orig_vol = pe.Node(
+        afni.Calc(expr="a*step(b)", outputtype="NIFTI_GZ"), name="sstrip_orig_vol"
+    )
+    binarize = pe.Node(Binarize(thresh_low=0.0), name="binarize")
 
     if unifize:
         # Add two unifize steps, pre- and post- skullstripping.
-        inu_uni_0 = pe.Node(afni.Unifize(outputtype='NIFTI_GZ'),
-                            name='unifize_pre_skullstrip')
-        inu_uni_1 = pe.Node(afni.Unifize(gm=True, outputtype='NIFTI_GZ'),
-                            name='unifize_post_skullstrip')
+        inu_uni_0 = pe.Node(
+            afni.Unifize(outputtype="NIFTI_GZ"), name="unifize_pre_skullstrip"
+        )
+        inu_uni_1 = pe.Node(
+            afni.Unifize(gm=True, outputtype="NIFTI_GZ"), name="unifize_post_skullstrip"
+        )
+        # fmt: off
         workflow.connect([
-            (inu_n4, inu_uni_0, [('output_image', 'in_file')]),
-            (inu_uni_0, sstrip, [('out_file', 'in_file')]),
-            (inu_uni_0, sstrip_orig_vol, [('out_file', 'in_file_a')]),
-            (sstrip_orig_vol, inu_uni_1, [('out_file', 'in_file')]),
-            (inu_uni_1, outputnode, [('out_file', 'out_file')]),
-            (inu_uni_0, outputnode, [('out_file', 'bias_corrected')]),
+            (inu_n4, inu_uni_0, [("output_image", "in_file")]),
+            (inu_uni_0, sstrip, [("out_file", "in_file")]),
+            (inu_uni_0, sstrip_orig_vol, [("out_file", "in_file_a")]),
+            (sstrip_orig_vol, inu_uni_1, [("out_file", "in_file")]),
+            (inu_uni_1, outputnode, [("out_file", "out_file")]),
+            (inu_uni_0, outputnode, [("out_file", "bias_corrected")]),
         ])
+        # fmt: on
     else:
+        # fmt: off
         workflow.connect([
-            (inputnode, sstrip_orig_vol, [('in_file', 'in_file_a')]),
-            (inu_n4, sstrip, [('output_image', 'in_file')]),
-            (sstrip_orig_vol, outputnode, [('out_file', 'out_file')]),
-            (inu_n4, outputnode, [('output_image', 'bias_corrected')]),
+            (inputnode, sstrip_orig_vol, [("in_file", "in_file_a")]),
+            (inu_n4, sstrip, [("output_image", "in_file")]),
+            (sstrip_orig_vol, outputnode, [("out_file", "out_file")]),
+            (inu_n4, outputnode, [("output_image", "bias_corrected")]),
         ])
+        # fmt: on
 
     # Remaining connections
+    # fmt: off
     workflow.connect([
-        (sstrip, sstrip_orig_vol, [('out_file', 'in_file_b')]),
-        (inputnode, inu_n4, [('in_file', 'input_image')]),
-        (sstrip_orig_vol, binarize, [('out_file', 'in_file')]),
-        (binarize, outputnode, [('out_mask', 'out_mask')]),
-        (inu_n4, outputnode, [('bias_image', 'bias_image')]),
+        (sstrip, sstrip_orig_vol, [("out_file", "in_file_b")]),
+        (inputnode, inu_n4, [("in_file", "input_image")]),
+        (sstrip_orig_vol, binarize, [("out_file", "in_file")]),
+        (binarize, outputnode, [("out_mask", "out_mask")]),
+        (inu_n4, outputnode, [("bias_image", "bias_image")]),
     ])
+    # fmt: on
     return workflow
