@@ -27,12 +27,14 @@ class FixHeaderApplyTransforms(ApplyTransforms):
     def _run_interface(self, runtime, correct_return_codes=(0,)):
         # Run normally
         runtime = super(FixHeaderApplyTransforms, self)._run_interface(
-            runtime, correct_return_codes)
+            runtime, correct_return_codes
+        )
 
-        _copyxform(self.inputs.reference_image,
-                   os.path.abspath(self._gen_filename('output_image')),
-                   message='%s (niworkflows v%s)' % (
-                       self.__class__.__name__, __version__))
+        _copyxform(
+            self.inputs.reference_image,
+            os.path.abspath(self._gen_filename("output_image")),
+            message="%s (niworkflows v%s)" % (self.__class__.__name__, __version__),
+        )
         return runtime
 
 
@@ -46,35 +48,42 @@ class FixHeaderRegistration(Registration):
     def _run_interface(self, runtime, correct_return_codes=(0,)):
         # Run normally
         runtime = super(FixHeaderRegistration, self)._run_interface(
-            runtime, correct_return_codes)
+            runtime, correct_return_codes
+        )
 
         # Forward transform
         out_file = self._get_outputfilenames(inverse=False)
         if out_file is not None and out_file:
             _copyxform(
-                self.inputs.fixed_image[0], os.path.abspath(out_file),
-                message='%s (niworkflows v%s)' % (
-                    self.__class__.__name__, __version__))
+                self.inputs.fixed_image[0],
+                os.path.abspath(out_file),
+                message="%s (niworkflows v%s)" % (self.__class__.__name__, __version__),
+            )
 
         # Inverse transform
         out_file = self._get_outputfilenames(inverse=True)
         if out_file is not None and out_file:
             _copyxform(
-                self.inputs.moving_image[0], os.path.abspath(out_file),
-                message='%s (niworkflows v%s)' % (
-                    self.__class__.__name__, __version__))
+                self.inputs.moving_image[0],
+                os.path.abspath(out_file),
+                message="%s (niworkflows v%s)" % (self.__class__.__name__, __version__),
+            )
 
         return runtime
 
 
 class _FixN4BiasFieldCorrectionOutputSpec(VanillaN4OutputSpec):
-    negative_values = traits.Bool(False, usedefault=True,
-                                  desc='Indicates whether the input was corrected for '
-                                       'nonpositive values by adding a constant offset.')
+    negative_values = traits.Bool(
+        False,
+        usedefault=True,
+        desc="Indicates whether the input was corrected for "
+        "nonpositive values by adding a constant offset.",
+    )
 
 
 class FixN4BiasFieldCorrection(VanillaN4):
     """Checks and fixes for nonpositive values in the input to ``N4BiasFieldCorrection``."""
+
     output_spec = _FixN4BiasFieldCorrectionOutputSpec
 
     def __init__(self, *args, **kwargs):
@@ -84,10 +93,11 @@ class FixN4BiasFieldCorrection(VanillaN4):
         super(FixN4BiasFieldCorrection, self).__init__(*args, **kwargs)
 
     def _format_arg(self, name, trait_spec, value):
-        if name == 'input_image':
+        if name == "input_image":
             return trait_spec.argstr % self._input_image
         return super(FixN4BiasFieldCorrection, self)._format_arg(
-            name, trait_spec, value)
+            name, trait_spec, value
+        )
 
     def _parse_inputs(self, skip=None):
         self._input_image = self.inputs.input_image
@@ -95,8 +105,9 @@ class FixN4BiasFieldCorrection(VanillaN4):
         input_nii = nb.load(self.inputs.input_image)
         datamin = input_nii.get_fdata().min()
         if datamin < 0:
-            self._input_image = fname_presuffix(self.inputs.input_image, suffix='_scaled',
-                                                newpath=os.getcwd())
+            self._input_image = fname_presuffix(
+                self.inputs.input_image, suffix="_scaled", newpath=os.getcwd()
+            )
             data = input_nii.get_fdata() - datamin
             newnii = input_nii.__class__(data, input_nii.affine, input_nii.header)
             newnii.to_filename(self._input_image)
@@ -106,5 +117,5 @@ class FixN4BiasFieldCorrection(VanillaN4):
 
     def _list_outputs(self):
         outputs = super(FixN4BiasFieldCorrection, self)._list_outputs()
-        outputs['negative_values'] = self._negative_values
+        outputs["negative_values"] = self._negative_values
         return outputs
