@@ -11,17 +11,67 @@ import transforms3d
 from nipype import logging
 from nipype.utils.filemanip import fname_presuffix
 from nipype.interfaces.base import (
-    traits, TraitedSpec, BaseInterfaceInputSpec, SimpleInterface,
-    File, InputMultiPath, OutputMultiPath, isdefined)
+    traits,
+    TraitedSpec,
+    BaseInterfaceInputSpec,
+    SimpleInterface,
+    File,
+    InputMultiPath,
+    OutputMultiPath,
+    isdefined,
+)
 from nipype.interfaces import fsl
 
-LOGGER = logging.getLogger('nipype.interface')
+LOGGER = logging.getLogger("nipype.interface")
+
+
+class _RegridToZoomsInputSpec(BaseInterfaceInputSpec):
+    in_file = File(
+        exists=True, mandatory=True, desc="a file whose resolution is to change"
+    )
+    zooms = traits.Tuple(
+        traits.Float,
+        traits.Float,
+        traits.Float,
+        mandatory=True,
+        desc="the new resolution",
+    )
+    order = traits.Int(3, usedefault=True, desc="order of interpolator")
+    clip = traits.Bool(
+        True,
+        usedefault=True,
+        desc="clip the data array within the original image's range",
+    )
+
+
+class _RegridToZoomsOutputSpec(TraitedSpec):
+    out_file = File(exists=True, dec="the regridded file")
+
+
+class RegridToZooms(SimpleInterface):
+    """Change the resolution of an image (regrid)."""
+
+    input_spec = _RegridToZoomsInputSpec
+    output_spec = _RegridToZoomsOutputSpec
+
+    def _run_interface(self, runtime):
+        from ..utils.images import resample_by_spacing
+
+        self._results["out_file"] = fname_presuffix(
+            self.inputs.in_file, suffix="_regrid", newpath=runtime.cwd
+        )
+        resample_by_spacing(
+            self.inputs.in_file,
+            self.inputs.zooms,
+            order=self.inputs.order,
+            clip=self.inputs.clip,
+        ).to_filename(self._results["out_file"])
+        return runtime
 
 
 class _IntraModalMergeInputSpec(BaseInterfaceInputSpec):
-    in_files = InputMultiPath(File(exists=True), mandatory=True,
-                              desc='input files')
-    in_mask = File(exists=True, desc='input mask for grand mean scaling')
+    in_files = InputMultiPath(File(exists=True), mandatory=True, desc="input files")
+    in_mask = File(exists=True, desc="input mask for grand mean scaling")
     hmc = traits.Bool(True, usedefault=True)
     zero_based_avg = traits.Bool(True, usedefault=True)
     to_ras = traits.Bool(True, usedefault=True)
@@ -29,10 +79,10 @@ class _IntraModalMergeInputSpec(BaseInterfaceInputSpec):
 
 
 class _IntraModalMergeOutputSpec(TraitedSpec):
-    out_file = File(exists=True, desc='merged image')
-    out_avg = File(exists=True, desc='average image')
-    out_mats = OutputMultiPath(File(exists=True), desc='output matrices')
-    out_movpar = OutputMultiPath(File(exists=True), desc='output movement parameters')
+    out_file = File(exists=True, desc="merged image")
+    out_avg = File(exists=True, desc="average image")
+    out_mats = OutputMultiPath(File(exists=True), desc="output matrices")
+    out_movpar = OutputMultiPath(File(exists=True), desc="output movement parameters")
 
 
 class IntraModalMerge(SimpleInterface):
@@ -53,8 +103,7 @@ class IntraModalMerge(SimpleInterface):
             in_files = [self.inputs.in_files]
 
         if self.inputs.to_ras:
-            in_files = [reorient(inf, newpath=runtime.cwd)
-                        for inf in in_files]
+            in_files = [reorient(inf, newpath=runtime.cwd) for inf in in_files]
 
         run_hmc = self.inputs.hmc and len(in_files) > 1
 
@@ -64,7 +113,7 @@ class IntraModalMerge(SimpleInterface):
             filenii = nb.load(f)
             filenii = nb.squeeze_image(filenii)
             if len(filenii.shape) == 5:
-                raise RuntimeError('Input image (%s) is 5D.' % f)
+                raise RuntimeError("Input image (%s) is 5D." % f)
             if filenii.dataobj.ndim == 4:
                 nii_list += nb.four_to_three(filenii)
             else:
@@ -75,37 +124,44 @@ class IntraModalMerge(SimpleInterface):
         else:
             filenii = nii_list[0]
 
-        merged_fname = fname_presuffix(self.inputs.in_files[0],
-                                       suffix='_merged', newpath=runtime.cwd)
+        merged_fname = fname_presuffix(
+            self.inputs.in_files[0], suffix="_merged", newpath=runtime.cwd
+        )
         filenii.to_filename(merged_fname)
-        self._results['out_file'] = merged_fname
-        self._results['out_avg'] = merged_fname
+        self._results["out_file"] = merged_fname
+        self._results["out_avg"] = merged_fname
 
         if filenii.dataobj.ndim < 4:
             # TODO: generate identity out_mats and zero-filled out_movpar
             return runtime
 
         if run_hmc:
-            mcflirt = fsl.MCFLIRT(cost='normcorr', save_mats=True, save_plots=True,
-                                  ref_vol=0, in_file=merged_fname)
+            mcflirt = fsl.MCFLIRT(
+                cost="normcorr",
+                save_mats=True,
+                save_plots=True,
+                ref_vol=0,
+                in_file=merged_fname,
+            )
             mcres = mcflirt.run()
             filenii = nb.load(mcres.outputs.out_file)
-            self._results['out_file'] = mcres.outputs.out_file
-            self._results['out_mats'] = mcres.outputs.mat_file
-            self._results['out_movpar'] = mcres.outputs.par_file
+            self._results["out_file"] = mcres.outputs.out_file
+            self._results["out_mats"] = mcres.outputs.mat_file
+            self._results["out_movpar"] = mcres.outputs.par_file
 
-        hmcdata = filenii.get_fdata(dtype='float32')
+        hmcdata = filenii.get_fdata(dtype="float32")
         if self.inputs.grand_mean_scaling:
             if not isdefined(self.inputs.in_mask):
                 mean = np.median(hmcdata, axis=-1)
                 thres = np.percentile(mean, 25)
                 mask = mean > thres
             else:
-                mask = nb.load(self.inputs.in_mask).get_fdata(dtype='float32') > 0.5
+                mask = nb.load(self.inputs.in_mask).get_fdata(dtype="float32") > 0.5
 
             nimgs = hmcdata.shape[-1]
-            means = np.median(hmcdata[mask[..., np.newaxis]].reshape((-1, nimgs)).T,
-                              axis=-1)
+            means = np.median(
+                hmcdata[mask[..., np.newaxis]].reshape((-1, nimgs)).T, axis=-1
+            )
             max_mean = means.max()
             for i in range(nimgs):
                 hmcdata[..., i] *= max_mean / means[i]
@@ -114,11 +170,12 @@ class IntraModalMerge(SimpleInterface):
         if self.inputs.zero_based_avg:
             hmcdata -= hmcdata.min()
 
-        self._results['out_avg'] = fname_presuffix(self.inputs.in_files[0],
-                                                   suffix='_avg', newpath=runtime.cwd)
-        nb.Nifti1Image(
-            hmcdata, filenii.affine, filenii.header).to_filename(
-            self._results['out_avg'])
+        self._results["out_avg"] = fname_presuffix(
+            self.inputs.in_files[0], suffix="_avg", newpath=runtime.cwd
+        )
+        nb.Nifti1Image(hmcdata, filenii.affine, filenii.header).to_filename(
+            self._results["out_avg"]
+        )
 
         return runtime
 
@@ -138,18 +195,23 @@ DISCARD_TEMPLATE = """\t\t\t\t<li><abbr title="{path}">{basename}</abbr></li>"""
 
 
 class _TemplateDimensionsInputSpec(BaseInterfaceInputSpec):
-    t1w_list = InputMultiPath(File(exists=True), mandatory=True, desc='input T1w images')
-    max_scale = traits.Float(3.0, usedefault=True,
-                             desc='Maximum scaling factor in images to accept')
+    t1w_list = InputMultiPath(
+        File(exists=True), mandatory=True, desc="input T1w images"
+    )
+    max_scale = traits.Float(
+        3.0, usedefault=True, desc="Maximum scaling factor in images to accept"
+    )
 
 
 class _TemplateDimensionsOutputSpec(TraitedSpec):
-    t1w_valid_list = OutputMultiPath(exists=True, desc='valid T1w images')
-    target_zooms = traits.Tuple(traits.Float, traits.Float, traits.Float,
-                                desc='Target zoom information')
-    target_shape = traits.Tuple(traits.Int, traits.Int, traits.Int,
-                                desc='Target shape information')
-    out_report = File(exists=True, desc='conformation report')
+    t1w_valid_list = OutputMultiPath(exists=True, desc="valid T1w images")
+    target_zooms = traits.Tuple(
+        traits.Float, traits.Float, traits.Float, desc="Target zoom information"
+    )
+    target_shape = traits.Tuple(
+        traits.Int, traits.Int, traits.Int, desc="Target shape information"
+    )
+    out_report = File(exists=True, desc="conformation report")
 
 
 class TemplateDimensions(SimpleInterface):
@@ -167,19 +229,26 @@ class TemplateDimensions(SimpleInterface):
     To select images that require no scaling (i.e. all have smallest voxel sizes),
     set ``max_scale=1``.
     """
+
     input_spec = _TemplateDimensionsInputSpec
     output_spec = _TemplateDimensionsOutputSpec
 
     def _generate_segment(self, discards, dims, zooms):
-        items = [DISCARD_TEMPLATE.format(path=path, basename=os.path.basename(path))
-                 for path in discards]
-        discard_list = '\n'.join(["\t\t\t<ul>"] + items + ['\t\t\t</ul>']) if items else ''
-        zoom_fmt = '{:.02g}mm x {:.02g}mm x {:.02g}mm'.format(*zooms)
-        return CONFORMATION_TEMPLATE.format(n_t1w=len(self.inputs.t1w_list),
-                                            dims='x'.join(map(str, dims)),
-                                            zooms=zoom_fmt,
-                                            n_discards=len(discards),
-                                            discard_list=discard_list)
+        items = [
+            DISCARD_TEMPLATE.format(path=path, basename=os.path.basename(path))
+            for path in discards
+        ]
+        discard_list = (
+            "\n".join(["\t\t\t<ul>"] + items + ["\t\t\t</ul>"]) if items else ""
+        )
+        zoom_fmt = "{:.02g}mm x {:.02g}mm x {:.02g}mm".format(*zooms)
+        return CONFORMATION_TEMPLATE.format(
+            n_t1w=len(self.inputs.t1w_list),
+            dims="x".join(map(str, dims)),
+            zooms=zoom_fmt,
+            n_discards=len(discards),
+            discard_list=discard_list,
+        )
 
     def _run_interface(self, runtime):
         # Load images, orient as RAS, collect shape and zoom data
@@ -200,38 +269,40 @@ class TemplateDimensions(SimpleInterface):
 
         # Ignore dropped images
         valid_fnames = np.atleast_1d(in_names[valid]).tolist()
-        self._results['t1w_valid_list'] = valid_fnames
+        self._results["t1w_valid_list"] = valid_fnames
 
         # Set target shape information
         target_zooms = all_zooms[valid].min(axis=0)
         target_shape = all_shapes[valid].max(axis=0)
 
-        self._results['target_zooms'] = tuple(target_zooms.tolist())
-        self._results['target_shape'] = tuple(target_shape.tolist())
+        self._results["target_zooms"] = tuple(target_zooms.tolist())
+        self._results["target_shape"] = tuple(target_shape.tolist())
 
         # Create report
         dropped_images = in_names[~valid]
         segment = self._generate_segment(dropped_images, target_shape, target_zooms)
-        out_report = os.path.join(runtime.cwd, 'report.html')
-        with open(out_report, 'w') as fobj:
+        out_report = os.path.join(runtime.cwd, "report.html")
+        with open(out_report, "w") as fobj:
             fobj.write(segment)
 
-        self._results['out_report'] = out_report
+        self._results["out_report"] = out_report
 
         return runtime
 
 
 class _ConformInputSpec(BaseInterfaceInputSpec):
-    in_file = File(exists=True, mandatory=True, desc='Input image')
-    target_zooms = traits.Tuple(traits.Float, traits.Float, traits.Float,
-                                desc='Target zoom information')
-    target_shape = traits.Tuple(traits.Int, traits.Int, traits.Int,
-                                desc='Target shape information')
+    in_file = File(exists=True, mandatory=True, desc="Input image")
+    target_zooms = traits.Tuple(
+        traits.Float, traits.Float, traits.Float, desc="Target zoom information"
+    )
+    target_shape = traits.Tuple(
+        traits.Int, traits.Int, traits.Int, desc="Target shape information"
+    )
 
 
 class _ConformOutputSpec(TraitedSpec):
-    out_file = File(exists=True, desc='Conformed image')
-    transform = File(exists=True, desc='Conformation transform (voxel-to-voxel)')
+    out_file = File(exists=True, desc="Conformed image")
+    transform = File(exists=True, desc="Conformation transform (voxel-to-voxel)")
 
 
 class Conform(SimpleInterface):
@@ -246,6 +317,7 @@ class Conform(SimpleInterface):
     transform is the identity transform.
 
     """
+
     input_spec = _ConformInputSpec
     output_spec = _ConformOutputSpec
 
@@ -265,66 +337,80 @@ class Conform(SimpleInterface):
 
         # Reconstruct transform from orig to reoriented image
         ornt_xfm = nb.orientations.inv_ornt_aff(
-            nb.io_orientation(orig_img.affine), orig_img.shape)
+            nb.io_orientation(orig_img.affine), orig_img.shape
+        )
         # Identity unless proven otherwise
         target_affine = reoriented.affine.copy()
         conform_xfm = np.eye(4)
 
         xyz_unit = reoriented.header.get_xyzt_units()[0]
-        if xyz_unit == 'unknown':
+        if xyz_unit == "unknown":
             # Common assumption; if we're wrong, unlikely to be the only thing that breaks
-            xyz_unit = 'mm'
+            xyz_unit = "mm"
 
         # Set a 0.05mm threshold to performing rescaling
-        atol = {'meter': 1e-5, 'mm': 0.01, 'micron': 10}[xyz_unit]
+        atol_gross = {"meter": 5e-5, "mm": 0.05, "micron": 50}[xyz_unit]
+        # if 0.01 > difference > 0.001mm, freesurfer won't be able to merge the images
+        atol_fine = {"meter": 1e-6, "mm": 0.001, "micron": 1}[xyz_unit]
 
-        # Rescale => change zooms
-        # Resize => update image dimensions
-        rescale = not np.allclose(zooms, target_zooms, atol=atol)
+        # Update zooms => Modify affine
+        # Rescale => Resample to resized voxels
+        # Resize => Resample to new image dimensions
+        update_zooms = not np.allclose(zooms, target_zooms, atol=atol_fine, rtol=0)
+        rescale = not np.allclose(zooms, target_zooms, atol=atol_gross, rtol=0)
         resize = not np.all(shape == target_shape)
-        if rescale or resize:
-            if rescale:
+        resample = rescale or resize
+        if resample or update_zooms:
+            # Use an affine with the corrected zooms, whether or not we resample
+            if update_zooms:
                 scale_factor = target_zooms / zooms
-                target_affine[:3, :3] = reoriented.affine[:3, :3].dot(np.diag(scale_factor))
+                target_affine[:3, :3] = reoriented.affine[:3, :3] @ np.diag(scale_factor)
 
             if resize:
                 # The shift is applied after scaling.
                 # Use a proportional shift to maintain relative position in dataset
                 size_factor = target_span / (zooms * shape)
                 # Use integer shifts to avoid unnecessary interpolation
-                offset = (reoriented.affine[:3, 3] * size_factor - reoriented.affine[:3, 3])
+                offset = reoriented.affine[:3, 3] * size_factor - reoriented.affine[:3, 3]
                 target_affine[:3, 3] = reoriented.affine[:3, 3] + offset.astype(int)
 
-            data = nli.resample_img(reoriented, target_affine, target_shape).dataobj
-            conform_xfm = np.linalg.inv(reoriented.affine).dot(target_affine)
+            conform_xfm = np.linalg.inv(reoriented.affine) @ target_affine
+
+            # Create new image
+            data = reoriented.dataobj
+            if resample:
+                data = nli.resample_img(reoriented, target_affine, target_shape).dataobj
             reoriented = reoriented.__class__(data, target_affine, reoriented.header)
 
         # Image may be reoriented, rescaled, and/or resized
         if reoriented is not orig_img:
-            out_name = fname_presuffix(fname, suffix='_ras', newpath=runtime.cwd)
+            out_name = fname_presuffix(fname, suffix="_ras", newpath=runtime.cwd)
             reoriented.to_filename(out_name)
         else:
             out_name = fname
 
         transform = ornt_xfm.dot(conform_xfm)
-        assert np.allclose(orig_img.affine.dot(transform), target_affine)
+        if not np.allclose(orig_img.affine.dot(transform), target_affine):
+            raise ValueError("Original and target affines are not similar")
 
-        mat_name = fname_presuffix(fname, suffix='.mat', newpath=runtime.cwd, use_ext=False)
-        np.savetxt(mat_name, transform, fmt='%.08f')
+        mat_name = fname_presuffix(
+            fname, suffix=".mat", newpath=runtime.cwd, use_ext=False
+        )
+        np.savetxt(mat_name, transform, fmt="%.08f")
 
-        self._results['out_file'] = out_name
-        self._results['transform'] = mat_name
+        self._results["out_file"] = out_name
+        self._results["transform"] = mat_name
 
         return runtime
 
 
 class _ValidateImageInputSpec(BaseInterfaceInputSpec):
-    in_file = File(exists=True, mandatory=True, desc='input image')
+    in_file = File(exists=True, mandatory=True, desc="input image")
 
 
 class _ValidateImageOutputSpec(TraitedSpec):
-    out_file = File(exists=True, desc='validated image')
-    out_report = File(exists=True, desc='HTML segment containing warning')
+    out_file = File(exists=True, desc="validated image")
+    out_report = File(exists=True, desc="HTML segment containing warning")
 
 
 class ValidateImage(SimpleInterface):
@@ -332,7 +418,7 @@ class ValidateImage(SimpleInterface):
     Check the correctness of x-form headers (matrix and code).
 
     This interface implements the `following logic
-    <https://github.com/poldracklab/fmriprep/issues/873#issuecomment-349394544>`_:
+    <https://github.com/nipreps/fmriprep/issues/873#issuecomment-349394544>`_:
 
     +-------------------+------------------+------------------+------------------\
 +------------------------------------------------+
@@ -372,11 +458,11 @@ class ValidateImage(SimpleInterface):
 
     def _run_interface(self, runtime):
         img = nb.load(self.inputs.in_file)
-        out_report = os.path.join(runtime.cwd, 'report.html')
+        out_report = os.path.join(runtime.cwd, "report.html")
 
         # Retrieve xform codes
-        sform_code = int(img.header._structarr['sform_code'])
-        qform_code = int(img.header._structarr['qform_code'])
+        sform_code = int(img.header._structarr["sform_code"])
+        qform_code = int(img.header._structarr["qform_code"])
 
         # Check qform is valid
         valid_qform = False
@@ -399,47 +485,58 @@ class ValidateImage(SimpleInterface):
 
         # Both match, qform valid (implicit with match), codes okay -> do nothing, empty report
         if matching_affines and qform_code > 0 and sform_code > 0:
-            self._results['out_file'] = self.inputs.in_file
-            open(out_report, 'w').close()
-            self._results['out_report'] = out_report
+            self._results["out_file"] = self.inputs.in_file
+            open(out_report, "w").close()
+            self._results["out_report"] = out_report
             return runtime
 
         # A new file will be written
-        out_fname = fname_presuffix(self.inputs.in_file, suffix='_valid', newpath=runtime.cwd)
-        self._results['out_file'] = out_fname
+        out_fname = fname_presuffix(
+            self.inputs.in_file, suffix="_valid", newpath=runtime.cwd
+        )
+        self._results["out_file"] = out_fname
 
         # Row 2:
         if valid_qform and qform_code > 0 and (sform_code == 0 or not valid_sform):
             img.set_sform(qform, qform_code)
-            warning_txt = 'Note on orientation: sform matrix set'
+            warning_txt = "Note on orientation: sform matrix set"
             description = """\
 <p class="elem-desc">The sform has been copied from qform.</p>
 """
         # Rows 3-4:
         # Note: if qform is not valid, matching_affines is False
-        elif (valid_sform and sform_code > 0) and (not matching_affines or qform_code == 0):
+        elif (valid_sform and sform_code > 0) and (
+            not matching_affines or qform_code == 0
+        ):
             img.set_qform(sform, sform_code)
             new_qform = img.get_qform()
-            if np.allclose(new_qform, qform) and qform_code > 0:
-                # False alarm
-                self._results['out_file'] = self.inputs.in_file
-                open(out_report, 'w').close()
-                self._results['out_report'] = out_report
-                return runtime
-            diff = np.linalg.inv(qform) @ new_qform
-            trans, rot, _, _ = transforms3d.affines.decompose44(diff)
-            angle = transforms3d.axangles.mat2axangle(rot)[1]
-            total_trans = np.sqrt(np.sum(trans * trans))  # Add angle and total_trans to report
-            warning_txt = 'Note on orientation: qform matrix overwritten'
-            description = """\
+            if valid_qform:
+                # False alarm - the difference is due to precision loss of qform
+                if np.allclose(new_qform, qform) and qform_code > 0:
+                    self._results["out_file"] = self.inputs.in_file
+                    open(out_report, "w").close()
+                    self._results["out_report"] = out_report
+                    return runtime
+                # Replacing an existing, valid qform. Report magnitude of change.
+                diff = np.linalg.inv(qform) @ new_qform
+                trans, rot, _, _ = transforms3d.affines.decompose44(diff)
+                angle = transforms3d.axangles.mat2axangle(rot)[1]
+                total_trans = np.sqrt(
+                    np.sum(trans * trans)
+                )  # Add angle and total_trans to report
+                warning_txt = "Note on orientation: qform matrix overwritten"
+                description = """\
     <p class="elem-desc">
     The qform has been copied from sform.
     The difference in angle is {angle:.02g}.
     The difference in translation is {total_trans:.02g}.
     </p>
-    """.format(angle=angle, total_trans=total_trans)
-            if not valid_qform and qform_code > 0:
-                warning_txt = 'WARNING - Invalid qform information'
+    """.format(
+                    angle=angle, total_trans=total_trans
+                )
+            elif qform_code > 0:
+                # qform code indicates the qform is supposed to be valid. Use more stridency.
+                warning_txt = "WARNING - Invalid qform information"
                 description = """\
 <p class="elem-desc">
     The qform matrix found in the file header is invalid.
@@ -448,12 +545,18 @@ class ValidateImage(SimpleInterface):
     by the scanner is advised.
 </p>
 """
+            else:  # qform_code == 0
+                # qform is not expected to be valids. Simple note.
+                warning_txt = "Note on orientation: qform matrix overwritten"
+                description = (
+                    '<p class="elem-desc">The qform has been copied from sform.</p>'
+                )
         # Rows 5-6:
         else:
             affine = img.header.get_base_affine()
-            img.set_sform(affine, nb.nifti1.xform_codes['scanner'])
-            img.set_qform(affine, nb.nifti1.xform_codes['scanner'])
-            warning_txt = 'WARNING - Missing orientation information'
+            img.set_sform(affine, nb.nifti1.xform_codes["scanner"])
+            img.set_qform(affine, nb.nifti1.xform_codes["scanner"])
+            warning_txt = "WARNING - Missing orientation information"
             description = """\
 <p class="elem-desc">
     FMRIPREP could not retrieve orientation information from the image header.
@@ -464,24 +567,23 @@ class ValidateImage(SimpleInterface):
         snippet = '<h3 class="elem-title">%s</h3>\n%s\n' % (warning_txt, description)
         # Store new file and report
         img.to_filename(out_fname)
-        with open(out_report, 'w') as fobj:
-            fobj.write(indent(snippet, '\t' * 3))
+        with open(out_report, "w") as fobj:
+            fobj.write(indent(snippet, "\t" * 3))
 
-        self._results['out_report'] = out_report
+        self._results["out_report"] = out_report
         return runtime
 
 
 class _DemeanImageInputSpec(BaseInterfaceInputSpec):
-    in_file = File(exists=True, mandatory=True,
-                   desc='image to be demeaned')
-    in_mask = File(exists=True, mandatory=True,
-                   desc='mask where median will be calculated')
-    only_mask = traits.Bool(False, usedefault=True,
-                            desc='demean only within mask')
+    in_file = File(exists=True, mandatory=True, desc="image to be demeaned")
+    in_mask = File(
+        exists=True, mandatory=True, desc="mask where median will be calculated"
+    )
+    only_mask = traits.Bool(False, usedefault=True, desc="demean only within mask")
 
 
 class _DemeanImageOutputSpec(TraitedSpec):
-    out_file = File(exists=True, desc='demeaned image')
+    out_file = File(exists=True, desc="demeaned image")
 
 
 class DemeanImage(SimpleInterface):
@@ -489,25 +591,25 @@ class DemeanImage(SimpleInterface):
     output_spec = _DemeanImageOutputSpec
 
     def _run_interface(self, runtime):
-        self._results['out_file'] = demean(
+        self._results["out_file"] = demean(
             self.inputs.in_file,
             self.inputs.in_mask,
             only_mask=self.inputs.only_mask,
-            newpath=runtime.cwd)
+            newpath=runtime.cwd,
+        )
         return runtime
 
 
 class _FilledImageLikeInputSpec(BaseInterfaceInputSpec):
-    in_file = File(exists=True, mandatory=True,
-                   desc='image to be demeaned')
-    fill_value = traits.Float(1.0, usedefault=True,
-                              desc='value to fill')
-    dtype = traits.Enum('float32', 'uint8', usedefault=True,
-                        desc='force output data type')
+    in_file = File(exists=True, mandatory=True, desc="image to be demeaned")
+    fill_value = traits.Float(1.0, usedefault=True, desc="value to fill")
+    dtype = traits.Enum(
+        "float32", "uint8", usedefault=True, desc="force output data type"
+    )
 
 
 class _FilledImageLikeOutputSpec(TraitedSpec):
-    out_file = File(exists=True, desc='demeaned image')
+    out_file = File(exists=True, desc="demeaned image")
 
 
 class FilledImageLike(SimpleInterface):
@@ -515,23 +617,26 @@ class FilledImageLike(SimpleInterface):
     output_spec = _FilledImageLikeOutputSpec
 
     def _run_interface(self, runtime):
-        self._results['out_file'] = nii_ones_like(
+        self._results["out_file"] = nii_ones_like(
             self.inputs.in_file,
             self.inputs.fill_value,
             self.inputs.dtype,
-            newpath=runtime.cwd)
+            newpath=runtime.cwd,
+        )
         return runtime
 
 
 class _MatchHeaderInputSpec(BaseInterfaceInputSpec):
-    reference = File(exists=True, mandatory=True,
-                     desc='NIfTI file with reference header')
-    in_file = File(exists=True, mandatory=True,
-                   desc='NIfTI file which header will be checked')
+    reference = File(
+        exists=True, mandatory=True, desc="NIfTI file with reference header"
+    )
+    in_file = File(
+        exists=True, mandatory=True, desc="NIfTI file which header will be checked"
+    )
 
 
 class _MatchHeaderOutputSpec(TraitedSpec):
-    out_file = File(exists=True, desc='NIfTI file with fixed header')
+    out_file = File(exists=True, desc="NIfTI file with fixed header")
 
 
 class MatchHeader(SimpleInterface):
@@ -543,53 +648,37 @@ class MatchHeader(SimpleInterface):
         imgnii = nb.load(self.inputs.in_file)
         imghdr = imgnii.header.copy()
 
-        imghdr['dim_info'] = refhdr['dim_info']  # dim_info is lost sometimes
+        imghdr["dim_info"] = refhdr["dim_info"]  # dim_info is lost sometimes
 
         # Set qform
         qform = refhdr.get_qform()
-        qcode = int(refhdr['qform_code'])
+        qcode = int(refhdr["qform_code"])
         if not np.allclose(qform, imghdr.get_qform()):
-            LOGGER.warning(
-                'q-forms of reference and mask are substantially different')
+            LOGGER.warning("q-forms of reference and mask are substantially different")
         imghdr.set_qform(qform, qcode)
 
         # Set sform
         sform = refhdr.get_sform()
-        scode = int(refhdr['sform_code'])
+        scode = int(refhdr["sform_code"])
         if not np.allclose(sform, imghdr.get_sform()):
-            LOGGER.warning(
-                's-forms of reference and mask are substantially different')
+            LOGGER.warning("s-forms of reference and mask are substantially different")
         imghdr.set_sform(sform, scode)
 
-        out_file = fname_presuffix(self.inputs.in_file, suffix='_hdr',
-                                   newpath=runtime.cwd)
+        out_file = fname_presuffix(
+            self.inputs.in_file, suffix="_hdr", newpath=runtime.cwd
+        )
 
-        imgnii.__class__(imgnii.dataobj, imghdr.get_best_affine(),
-                         imghdr).to_filename(out_file)
-        self._results['out_file'] = out_file
+        imgnii.__class__(imgnii.dataobj, imghdr.get_best_affine(), imghdr).to_filename(
+            out_file
+        )
+        self._results["out_file"] = out_file
         return runtime
 
 
 def reorient(in_file, newpath=None):
     """Reorient Nifti files to RAS."""
-    out_file = fname_presuffix(in_file, suffix='_ras', newpath=newpath)
+    out_file = fname_presuffix(in_file, suffix="_ras", newpath=newpath)
     nb.as_closest_canonical(nb.load(in_file)).to_filename(out_file)
-    return out_file
-
-
-def extract_wm(in_seg, wm_label=3, newpath=None):
-    import nibabel as nb
-    import numpy as np
-    from nipype.utils.filemanip import fname_presuffix
-
-    nii = nb.load(in_seg)
-    data = np.zeros(nii.shape, dtype=np.uint8)
-    data[np.asanyarray(nii.dataobj) == wm_label] = 1
-
-    out_file = fname_presuffix(in_seg, suffix='_wm', newpath=newpath)
-    new = nb.Nifti1Image(data, nii.affine, nii.header)
-    new.set_data_dtype(np.uint8)
-    new.to_filename(out_file)
     return out_file
 
 
@@ -616,9 +705,14 @@ def normalize_xform(img):
     # Check desired codes
     qform, qform_code = img.get_qform(coded=True)
     sform, sform_code = img.get_sform(coded=True)
-    if all((qform is not None and np.allclose(qform, xform),
+    if all(
+        (
+            qform is not None and np.allclose(qform, xform),
             sform is not None and np.allclose(sform, xform),
-            int(qform_code) == xform_code, int(sform_code) == xform_code)):
+            int(qform_code) == xform_code,
+            int(sform_code) == xform_code,
+        )
+    ):
         return img
 
     new_img = img.__class__(img.dataobj, xform, img.header)
@@ -635,8 +729,7 @@ def demean(in_file, in_mask, only_mask=False, newpath=None):
     import nibabel as nb
     from nipype.utils.filemanip import fname_presuffix
 
-    out_file = fname_presuffix(in_file, suffix='_demeaned',
-                               newpath=os.getcwd())
+    out_file = fname_presuffix(in_file, suffix="_demeaned", newpath=os.getcwd())
     nii = nb.load(in_file)
     msk = np.asanyarray(nb.load(in_mask).dataobj)
     data = nii.get_fdata()
@@ -644,8 +737,7 @@ def demean(in_file, in_mask, only_mask=False, newpath=None):
         data[msk > 0] -= np.median(data[msk > 0])
     else:
         data -= np.median(data[msk > 0])
-    nb.Nifti1Image(data, nii.affine, nii.header).to_filename(
-        out_file)
+    nb.Nifti1Image(data, nii.affine, nii.header).to_filename(out_file)
     return out_file
 
 
@@ -667,44 +759,48 @@ def nii_ones_like(in_file, value, dtype, newpath=None):
 
 
 class _SignalExtractionInputSpec(BaseInterfaceInputSpec):
-    in_file = File(exists=True, mandatory=True, desc='4-D fMRI nii file')
+    in_file = File(exists=True, mandatory=True, desc="4-D fMRI nii file")
     label_files = InputMultiPath(
         File(exists=True),
         mandatory=True,
-        desc='a 3D label image, with 0 denoting '
-        'background, or a list of 3D probability '
-        'maps (one per label) or the equivalent 4D '
-        'file.')
+        desc="a 3D label image, with 0 denoting "
+        "background, or a list of 3D probability "
+        "maps (one per label) or the equivalent 4D "
+        "file.",
+    )
     prob_thres = traits.Range(
         low=0.0,
         high=1.0,
         value=0.5,
         usedefault=True,
-        desc='If label_files are probability masks, threshold '
-        'at specified probability.')
+        desc="If label_files are probability masks, threshold "
+        "at specified probability.",
+    )
     class_labels = traits.List(
         mandatory=True,
-        desc='Human-readable labels for each segment '
-        'in the label file, in order. The length of '
-        'class_labels must be equal to the number of '
-        'segments (background excluded). This list '
-        'corresponds to the class labels in label_file '
-        'in ascending order')
+        desc="Human-readable labels for each segment "
+        "in the label file, in order. The length of "
+        "class_labels must be equal to the number of "
+        "segments (background excluded). This list "
+        "corresponds to the class labels in label_file "
+        "in ascending order",
+    )
     out_file = File(
-        'signals.tsv',
+        "signals.tsv",
         usedefault=True,
         exists=False,
-        desc='The name of the file to output to. '
-        'signals.tsv by default')
+        desc="The name of the file to output to. " "signals.tsv by default",
+    )
 
 
 class _SignalExtractionOutputSpec(TraitedSpec):
     out_file = File(
         exists=True,
-        desc='tsv file containing the computed '
-        'signals, with as many columns as there are labels and as '
-        'many rows as there are timepoints in in_file, plus a '
-        'header row with values from class_labels')
+        desc="tsv file containing the computed "
+        "signals, with as many columns as there are labels and as "
+        "many rows as there are timepoints in in_file, plus a "
+        "header row with values from class_labels",
+    )
 
 
 class SignalExtraction(SimpleInterface):
@@ -731,18 +827,20 @@ class SignalExtraction(SimpleInterface):
             raise NotImplementedError(
                 "Input image and mask should be of "
                 "same dimensions before running SignalExtraction"
-                )
+            )
         # Load the mask.
         # If mask is a list, each mask is treated as its own ROI/parcel
         # If mask is a 3D, each integer is treated as its own ROI/parcel
         if len(mask_imgs) > 1:
-            masks = [np.asanyarray(mask_img.dataobj) >= self.inputs.prob_thres
-                     for mask_img in mask_imgs]
+            masks = [
+                np.asanyarray(mask_img.dataobj) >= self.inputs.prob_thres
+                for mask_img in mask_imgs
+            ]
         else:
             labelsmap = np.asanyarray(mask_imgs[0].dataobj)
             labels = np.unique(labelsmap)
             labels = labels[labels != 0]
-            masks = [labelsmap == l for l in labels]
+            masks = [labelsmap == label for label in labels]
 
         if len(masks) != len(self.inputs.class_labels):
             raise ValueError("Number of masks must match number of labels")
@@ -754,9 +852,7 @@ class SignalExtraction(SimpleInterface):
             series[:, j] = data[mask, :].mean(axis=0)
 
         output = np.vstack((self.inputs.class_labels, series.astype(str)))
-        self._results['out_file'] = os.path.join(runtime.cwd,
-                                                 self.inputs.out_file)
-        np.savetxt(
-            self._results['out_file'], output, fmt=b'%s', delimiter='\t')
+        self._results["out_file"] = os.path.join(runtime.cwd, self.inputs.out_file)
+        np.savetxt(self._results["out_file"], output, fmt=b"%s", delimiter="\t")
 
         return runtime
