@@ -9,7 +9,13 @@ import scipy.ndimage as nd
 
 from nipype.interfaces import fsl, ants
 from nipype.interfaces.base import (
-    File, BaseInterfaceInputSpec, traits, isdefined, InputMultiPath, Str)
+    File,
+    BaseInterfaceInputSpec,
+    traits,
+    isdefined,
+    InputMultiPath,
+    Str,
+)
 from nipype.interfaces.mixins import reporting
 from nipype.algorithms import confounds
 from seaborn import color_palette
@@ -17,13 +23,13 @@ from .. import NIWORKFLOWS_LOG
 from . import report_base as nrc
 
 
-class _BETInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                       fsl.preprocess.BETInputSpec):
+class _BETInputSpecRPT(nrc._SVGReportCapableInputSpec, fsl.preprocess.BETInputSpec):
     pass
 
 
-class _BETOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                        fsl.preprocess.BETOutputSpec):
+class _BETOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, fsl.preprocess.BETOutputSpec
+):
     pass
 
 
@@ -38,27 +44,32 @@ class BETRPT(nrc.SegmentationRC, fsl.BET):
         return super(BETRPT, self)._run_interface(runtime)
 
     def _post_run_hook(self, runtime):
-        ''' generates a report showing slices from each axis of an arbitrary
-        volume of in_file, with the resulting binary brain mask overlaid '''
+        """ generates a report showing slices from each axis of an arbitrary
+        volume of in_file, with the resulting binary brain mask overlaid """
 
         self._anat_file = self.inputs.in_file
         self._mask_file = self.aggregate_outputs(runtime=runtime).mask_file
         self._seg_files = [self._mask_file]
         self._masked = self.inputs.mask
 
-        NIWORKFLOWS_LOG.info('Generating report for BET. file "%s", and mask file "%s"',
-                             self._anat_file, self._mask_file)
+        NIWORKFLOWS_LOG.info(
+            'Generating report for BET. file "%s", and mask file "%s"',
+            self._anat_file,
+            self._mask_file,
+        )
 
         return super(BETRPT, self)._post_run_hook(runtime)
 
 
-class _BrainExtractionInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                                   ants.segmentation.BrainExtractionInputSpec):
+class _BrainExtractionInputSpecRPT(
+    nrc._SVGReportCapableInputSpec, ants.segmentation.BrainExtractionInputSpec
+):
     pass
 
 
-class _BrainExtractionOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                                    ants.segmentation.BrainExtractionOutputSpec):
+class _BrainExtractionOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, ants.segmentation.BrainExtractionOutputSpec
+):
     pass
 
 
@@ -67,11 +78,16 @@ class BrainExtractionRPT(nrc.SegmentationRC, ants.segmentation.BrainExtraction):
     output_spec = _BrainExtractionOutputSpecRPT
 
     def _post_run_hook(self, runtime):
-        ''' generates a report showing slices from each axis '''
+        """ generates a report showing slices from each axis """
 
-        brain_extraction_mask = self.aggregate_outputs(runtime=runtime).BrainExtractionMask
+        brain_extraction_mask = self.aggregate_outputs(
+            runtime=runtime
+        ).BrainExtractionMask
 
-        if isdefined(self.inputs.keep_temporary_files) and self.inputs.keep_temporary_files == 1:
+        if (
+            isdefined(self.inputs.keep_temporary_files)
+            and self.inputs.keep_temporary_files == 1
+        ):
             self._anat_file = self.aggregate_outputs(runtime=runtime).N4Corrected0
         else:
             self._anat_file = self.inputs.anatomical_image
@@ -79,15 +95,17 @@ class BrainExtractionRPT(nrc.SegmentationRC, ants.segmentation.BrainExtraction):
         self._seg_files = [brain_extraction_mask]
         self._masked = False
 
-        NIWORKFLOWS_LOG.info('Generating report for ANTS BrainExtraction. file "%s", mask "%s"',
-                             self._anat_file, self._mask_file)
+        NIWORKFLOWS_LOG.info(
+            'Generating report for ANTS BrainExtraction. file "%s", mask "%s"',
+            self._anat_file,
+            self._mask_file,
+        )
 
         return super(BrainExtractionRPT, self)._post_run_hook(runtime)
 
 
 # TODO: move this interface to nipype.interfaces.nilearn
-class _ComputeEPIMaskInputSpec(nrc._SVGReportCapableInputSpec,
-                               BaseInterfaceInputSpec):
+class _ComputeEPIMaskInputSpec(nrc._SVGReportCapableInputSpec, BaseInterfaceInputSpec):
     in_file = File(exists=True, desc="3D or 4D EPI file")
     dilation = traits.Int(desc="binary dilation on the nilearn output")
 
@@ -105,11 +123,13 @@ class ComputeEPIMask(nrc.SegmentationRC):
         in_file_data = orig_file_nii.get_fdata()
 
         # pad the data to avoid the mask estimation running into edge effects
-        in_file_data_padded = np.pad(in_file_data, (1, 1), 'constant',
-                                     constant_values=(0, 0))
+        in_file_data_padded = np.pad(
+            in_file_data, (1, 1), "constant", constant_values=(0, 0)
+        )
 
-        padded_nii = nb.Nifti1Image(in_file_data_padded, orig_file_nii.affine,
-                                    orig_file_nii.header)
+        padded_nii = nb.Nifti1Image(
+            in_file_data_padded, orig_file_nii.affine, orig_file_nii.header
+        )
 
         mask_nii = compute_epi_mask(padded_nii, exclude_zeros=True)
 
@@ -124,8 +144,9 @@ class ComputeEPIMask(nrc.SegmentationRC):
         mask_data[in_file_data == 0] = 0
         mask_data[np.isnan(in_file_data)] = 0
 
-        better_mask = nb.Nifti1Image(mask_data, orig_file_nii.affine,
-                                     orig_file_nii.header)
+        better_mask = nb.Nifti1Image(
+            mask_data, orig_file_nii.affine, orig_file_nii.header
+        )
         better_mask.set_data_dtype(np.uint8)
         better_mask.to_filename("mask_file.nii.gz")
 
@@ -136,12 +157,12 @@ class ComputeEPIMask(nrc.SegmentationRC):
 
     def _list_outputs(self):
         outputs = super(ComputeEPIMask, self)._list_outputs()
-        outputs['mask_file'] = self._mask_file
+        outputs["mask_file"] = self._mask_file
         return outputs
 
     def _post_run_hook(self, runtime):
-        ''' generates a report showing slices from each axis of an arbitrary
-        volume of in_file, with the resulting binary brain mask overlaid '''
+        """ generates a report showing slices from each axis of an arbitrary
+        volume of in_file, with the resulting binary brain mask overlaid """
 
         self._anat_file = self.inputs.in_file
         self._mask_file = self.aggregate_outputs(runtime=runtime).mask_file
@@ -150,18 +171,20 @@ class ComputeEPIMask(nrc.SegmentationRC):
 
         NIWORKFLOWS_LOG.info(
             'Generating report for nilearn.compute_epi_mask. file "%s", and mask file "%s"',
-            self._anat_file, self._mask_file)
+            self._anat_file,
+            self._mask_file,
+        )
 
         return super(ComputeEPIMask, self)._post_run_hook(runtime)
 
 
-class _ACompCorInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                            confounds.CompCorInputSpec):
+class _ACompCorInputSpecRPT(nrc._SVGReportCapableInputSpec, confounds.CompCorInputSpec):
     pass
 
 
-class _ACompCorOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                             confounds.CompCorOutputSpec):
+class _ACompCorOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, confounds.CompCorOutputSpec
+):
     pass
 
 
@@ -170,29 +193,36 @@ class ACompCorRPT(nrc.SegmentationRC, confounds.ACompCor):
     output_spec = _ACompCorOutputSpecRPT
 
     def _post_run_hook(self, runtime):
-        ''' generates a report showing slices from each axis '''
+        """ generates a report showing slices from each axis """
 
-        assert len(self.inputs.mask_files) == 1, \
-            "ACompCorRPT only supports a single input mask. " \
-            "A list %s was found." % self.inputs.mask_files
+        if len(self.inputs.mask_files) != 1:
+            raise ValueError(
+                "ACompCorRPT only supports a single input mask. "
+                "A list %s was found." % self.inputs.mask_files
+            )
         self._anat_file = self.inputs.realigned_file
         self._mask_file = self.inputs.mask_files[0]
         self._seg_files = self.inputs.mask_files
         self._masked = False
 
-        NIWORKFLOWS_LOG.info('Generating report for aCompCor. file "%s", mask "%s"',
-                             self.inputs.realigned_file, self._mask_file)
+        NIWORKFLOWS_LOG.info(
+            'Generating report for aCompCor. file "%s", mask "%s"',
+            self.inputs.realigned_file,
+            self._mask_file,
+        )
 
         return super(ACompCorRPT, self)._post_run_hook(runtime)
 
 
-class _TCompCorInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                            confounds.TCompCorInputSpec):
+class _TCompCorInputSpecRPT(
+    nrc._SVGReportCapableInputSpec, confounds.TCompCorInputSpec
+):
     pass
 
 
-class _TCompCorOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                             confounds.TCompCorOutputSpec):
+class _TCompCorOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, confounds.TCompCorOutputSpec
+):
     pass
 
 
@@ -201,28 +231,34 @@ class TCompCorRPT(nrc.SegmentationRC, confounds.TCompCor):
     output_spec = _TCompCorOutputSpecRPT
 
     def _post_run_hook(self, runtime):
-        ''' generates a report showing slices from each axis '''
+        """ generates a report showing slices from each axis """
 
-        high_variance_masks = self.aggregate_outputs(runtime=runtime).high_variance_masks
+        high_variance_masks = self.aggregate_outputs(
+            runtime=runtime
+        ).high_variance_masks
 
-        assert not isinstance(high_variance_masks, list),\
-            "TCompCorRPT only supports a single output high variance mask. " \
-            "A list %s was found." % str(high_variance_masks)
+        if isinstance(high_variance_masks, list):
+            raise ValueError(
+                "TCompCorRPT only supports a single output high variance mask. "
+                "A list %s was found." % high_variance_masks
+            )
         self._anat_file = self.inputs.realigned_file
         self._mask_file = high_variance_masks
         self._seg_files = [high_variance_masks]
         self._masked = False
 
-        NIWORKFLOWS_LOG.info('Generating report for tCompCor. file "%s", mask "%s"',
-                             self.inputs.realigned_file,
-                             self.aggregate_outputs(runtime=runtime).high_variance_masks)
+        NIWORKFLOWS_LOG.info(
+            'Generating report for tCompCor. file "%s", mask "%s"',
+            self.inputs.realigned_file,
+            self.aggregate_outputs(runtime=runtime).high_variance_masks,
+        )
 
         return super(TCompCorRPT, self)._post_run_hook(runtime)
 
 
 class _SimpleShowMaskInputSpec(nrc._SVGReportCapableInputSpec):
-    background_file = File(exists=True, mandatory=True, desc='file before')
-    mask_file = File(exists=True, mandatory=True, desc='file before')
+    background_file = File(exists=True, mandatory=True, desc="file before")
+    mask_file = File(exists=True, mandatory=True, desc="file before")
 
 
 class SimpleShowMaskRPT(nrc.SegmentationRC, nrc.ReportingInterface):
@@ -238,16 +274,24 @@ class SimpleShowMaskRPT(nrc.SegmentationRC, nrc.ReportingInterface):
 
 
 class _ROIsPlotInputSpecRPT(nrc._SVGReportCapableInputSpec):
-    in_file = File(exists=True, mandatory=True, desc='the volume where ROIs are defined')
-    in_rois = InputMultiPath(File(exists=True), mandatory=True,
-                             desc='a list of regions to be plotted')
-    in_mask = File(exists=True, desc='a special region, eg. the brain mask')
-    masked = traits.Bool(False, usedefault=True, desc='mask in_file prior plotting')
-    colors = traits.Either(None, traits.List(Str), usedefault=True,
-                           desc='use specific colors for contours')
-    levels = traits.Either(None, traits.List(traits.Float),
-                           usedefault=True, desc='pass levels to nilearn.plotting')
-    mask_color = Str('r', usedefault=True, desc='color for mask')
+    in_file = File(
+        exists=True, mandatory=True, desc="the volume where ROIs are defined"
+    )
+    in_rois = InputMultiPath(
+        File(exists=True), mandatory=True, desc="a list of regions to be plotted"
+    )
+    in_mask = File(exists=True, desc="a special region, eg. the brain mask")
+    masked = traits.Bool(False, usedefault=True, desc="mask in_file prior plotting")
+    colors = traits.Either(
+        None, traits.List(Str), usedefault=True, desc="use specific colors for contours"
+    )
+    levels = traits.Either(
+        None,
+        traits.List(traits.Float),
+        usedefault=True,
+        desc="pass levels to nilearn.plotting",
+    )
+    mask_color = Str("r", usedefault=True, desc="color for mask")
 
 
 class ROIsPlot(nrc.ReportingInterface):
@@ -255,19 +299,20 @@ class ROIsPlot(nrc.ReportingInterface):
 
     def _generate_report(self):
         from niworkflows.viz.utils import plot_segs, compose_view
+
         seg_files = self.inputs.in_rois
-        mask_file = None if not isdefined(self.inputs.in_mask) \
-            else self.inputs.in_mask
+        mask_file = None if not isdefined(self.inputs.in_mask) else self.inputs.in_mask
 
         # Remove trait decoration and replace None with []
-        levels = [l for l in self.inputs.levels or []]
+        levels = [level for level in self.inputs.levels or []]
         colors = [c for c in self.inputs.colors or []]
 
         if len(seg_files) == 1:  # in_rois is a segmentation
             nsegs = len(levels)
             if nsegs == 0:
-                levels = np.unique(np.round(
-                    nb.load(seg_files[0]).get_fdata(dtype='float32')))
+                levels = np.unique(
+                    np.round(nb.load(seg_files[0]).get_fdata(dtype="float32"))
+                )
                 levels = (levels[levels > 0] - 0.5).tolist()
                 nsegs = len(levels)
 
@@ -303,5 +348,5 @@ class ROIsPlot(nrc.ReportingInterface):
                 compress=self.inputs.compress_report,
             ),
             fg_svgs=None,
-            out_file=self._out_report
+            out_file=self._out_report,
         )
