@@ -10,7 +10,14 @@ from nilearn import image as nli
 from nilearn.image import index_img
 from nipype.utils.filemanip import fname_presuffix
 from nipype.interfaces.base import (
-    traits, isdefined, TraitedSpec, BaseInterfaceInputSpec, File, SimpleInterface)
+    traits,
+    isdefined,
+    TraitedSpec,
+    BaseInterfaceInputSpec,
+    File,
+    SimpleInterface,
+    InputMultiObject,
+)
 from nipype.interfaces.mixins import reporting
 from nipype.interfaces import freesurfer as fs
 from nipype.interfaces import fsl, ants, afni
@@ -20,49 +27,57 @@ from . import report_base as nrc
 from .mni import (
     _RobustMNINormalizationInputSpec,
     _RobustMNINormalizationOutputSpec,
-    RobustMNINormalization
+    RobustMNINormalization,
 )
-from .fixes import (FixHeaderApplyTransforms as ApplyTransforms,
-                    FixHeaderRegistration as Registration)
+from .fixes import (
+    FixHeaderApplyTransforms as ApplyTransforms,
+    FixHeaderRegistration as Registration,
+)
 
 
 class _RobustMNINormalizationInputSpecRPT(
-        nrc._SVGReportCapableInputSpec, _RobustMNINormalizationInputSpec):
+    nrc._SVGReportCapableInputSpec, _RobustMNINormalizationInputSpec
+):
     pass
 
 
 class _RobustMNINormalizationOutputSpecRPT(
-        reporting.ReportCapableOutputSpec, _RobustMNINormalizationOutputSpec):
+    reporting.ReportCapableOutputSpec, _RobustMNINormalizationOutputSpec
+):
     pass
 
 
-class RobustMNINormalizationRPT(
-        nrc.RegistrationRC, RobustMNINormalization):
+class RobustMNINormalizationRPT(nrc.RegistrationRC, RobustMNINormalization):
     input_spec = _RobustMNINormalizationInputSpecRPT
     output_spec = _RobustMNINormalizationOutputSpecRPT
 
     def _post_run_hook(self, runtime):
         # We need to dig into the internal ants.Registration interface
-        self._fixed_image = self._get_ants_args()['fixed_image']
+        self._fixed_image = self._get_ants_args()["fixed_image"]
         if isinstance(self._fixed_image, (list, tuple)):
             self._fixed_image = self._fixed_image[0]  # get first item if list
 
-        if self._get_ants_args().get('fixed_image_mask') is not None:
-            self._fixed_image_mask = self._get_ants_args().get('fixed_image_mask')
+        if self._get_ants_args().get("fixed_image_mask") is not None:
+            self._fixed_image_mask = self._get_ants_args().get("fixed_image_mask")
         self._moving_image = self.aggregate_outputs(runtime=runtime).warped_image
-        NIWORKFLOWS_LOG.info('Report - setting fixed (%s) and moving (%s) images',
-                             self._fixed_image, self._moving_image)
+        NIWORKFLOWS_LOG.info(
+            "Report - setting fixed (%s) and moving (%s) images",
+            self._fixed_image,
+            self._moving_image,
+        )
 
         return super(RobustMNINormalizationRPT, self)._post_run_hook(runtime)
 
 
-class _ANTSRegistrationInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                                    ants.registration.RegistrationInputSpec):
+class _ANTSRegistrationInputSpecRPT(
+    nrc._SVGReportCapableInputSpec, ants.registration.RegistrationInputSpec
+):
     pass
 
 
-class _ANTSRegistrationOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                                     ants.registration.RegistrationOutputSpec):
+class _ANTSRegistrationOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, ants.registration.RegistrationOutputSpec
+):
     pass
 
 
@@ -73,19 +88,24 @@ class ANTSRegistrationRPT(nrc.RegistrationRC, Registration):
     def _post_run_hook(self, runtime):
         self._fixed_image = self.inputs.fixed_image[0]
         self._moving_image = self.aggregate_outputs(runtime=runtime).warped_image
-        NIWORKFLOWS_LOG.info('Report - setting fixed (%s) and moving (%s) images',
-                             self._fixed_image, self._moving_image)
+        NIWORKFLOWS_LOG.info(
+            "Report - setting fixed (%s) and moving (%s) images",
+            self._fixed_image,
+            self._moving_image,
+        )
 
         return super(ANTSRegistrationRPT, self)._post_run_hook(runtime)
 
 
-class _ANTSApplyTransformsInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                                       ants.resampling.ApplyTransformsInputSpec):
+class _ANTSApplyTransformsInputSpecRPT(
+    nrc._SVGReportCapableInputSpec, ants.resampling.ApplyTransformsInputSpec
+):
     pass
 
 
-class _ANTSApplyTransformsOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                                        ants.resampling.ApplyTransformsOutputSpec):
+class _ANTSApplyTransformsOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, ants.resampling.ApplyTransformsOutputSpec
+):
     pass
 
 
@@ -96,20 +116,24 @@ class ANTSApplyTransformsRPT(nrc.RegistrationRC, ApplyTransforms):
     def _post_run_hook(self, runtime):
         self._fixed_image = self.inputs.reference_image
         self._moving_image = self.aggregate_outputs(runtime=runtime).output_image
-        NIWORKFLOWS_LOG.info('Report - setting fixed (%s) and moving (%s) images',
-                             self._fixed_image, self._moving_image)
+        NIWORKFLOWS_LOG.info(
+            "Report - setting fixed (%s) and moving (%s) images",
+            self._fixed_image,
+            self._moving_image,
+        )
 
         return super(ANTSApplyTransformsRPT, self)._post_run_hook(runtime)
 
 
-class _ApplyTOPUPInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                              fsl.epi.ApplyTOPUPInputSpec):
-    wm_seg = File(argstr='-wmseg %s',
-                  desc='reference white matter segmentation mask')
+class _ApplyTOPUPInputSpecRPT(
+    nrc._SVGReportCapableInputSpec, fsl.epi.ApplyTOPUPInputSpec
+):
+    wm_seg = File(argstr="-wmseg %s", desc="reference white matter segmentation mask")
 
 
-class _ApplyTOPUPOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                               fsl.epi.ApplyTOPUPOutputSpec):
+class _ApplyTOPUPOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, fsl.epi.ApplyTOPUPOutputSpec
+):
     pass
 
 
@@ -120,23 +144,27 @@ class ApplyTOPUPRPT(nrc.RegistrationRC, fsl.ApplyTOPUP):
     def _post_run_hook(self, runtime):
         self._fixed_image_label = "after"
         self._moving_image_label = "before"
-        self._fixed_image = index_img(self.aggregate_outputs(runtime=runtime).out_corrected, 0)
+        self._fixed_image = index_img(
+            self.aggregate_outputs(runtime=runtime).out_corrected, 0
+        )
         self._moving_image = index_img(self.inputs.in_files[0], 0)
         self._contour = self.inputs.wm_seg if isdefined(self.inputs.wm_seg) else None
-        NIWORKFLOWS_LOG.info('Report - setting corrected (%s) and warped (%s) images',
-                             self._fixed_image, self._moving_image)
+        NIWORKFLOWS_LOG.info(
+            "Report - setting corrected (%s) and warped (%s) images",
+            self._fixed_image,
+            self._moving_image,
+        )
 
         return super(ApplyTOPUPRPT, self)._post_run_hook(runtime)
 
 
-class _FUGUEInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                         fsl.preprocess.FUGUEInputSpec):
-    wm_seg = File(argstr='-wmseg %s',
-                  desc='reference white matter segmentation mask')
+class _FUGUEInputSpecRPT(nrc._SVGReportCapableInputSpec, fsl.preprocess.FUGUEInputSpec):
+    wm_seg = File(argstr="-wmseg %s", desc="reference white matter segmentation mask")
 
 
-class _FUGUEOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                          fsl.preprocess.FUGUEOutputSpec):
+class _FUGUEOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, fsl.preprocess.FUGUEOutputSpec
+):
     pass
 
 
@@ -151,19 +179,21 @@ class FUGUERPT(nrc.RegistrationRC, fsl.FUGUE):
         self._moving_image = self.inputs.in_file
         self._contour = self.inputs.wm_seg if isdefined(self.inputs.wm_seg) else None
         NIWORKFLOWS_LOG.info(
-            'Report - setting corrected (%s) and warped (%s) images',
-            self._fixed_image, self._moving_image)
+            "Report - setting corrected (%s) and warped (%s) images",
+            self._fixed_image,
+            self._moving_image,
+        )
 
         return super(FUGUERPT, self)._post_run_hook(runtime)
 
 
-class _FLIRTInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                         fsl.preprocess.FLIRTInputSpec):
+class _FLIRTInputSpecRPT(nrc._SVGReportCapableInputSpec, fsl.preprocess.FLIRTInputSpec):
     pass
 
 
-class _FLIRTOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                          fsl.preprocess.FLIRTOutputSpec):
+class _FLIRTOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, fsl.preprocess.FLIRTOutputSpec
+):
     pass
 
 
@@ -176,14 +206,17 @@ class FLIRTRPT(nrc.RegistrationRC, fsl.FLIRT):
         self._moving_image = self.aggregate_outputs(runtime=runtime).out_file
         self._contour = self.inputs.wm_seg if isdefined(self.inputs.wm_seg) else None
         NIWORKFLOWS_LOG.info(
-            'Report - setting fixed (%s) and moving (%s) images',
-            self._fixed_image, self._moving_image)
+            "Report - setting fixed (%s) and moving (%s) images",
+            self._fixed_image,
+            self._moving_image,
+        )
 
         return super(FLIRTRPT, self)._post_run_hook(runtime)
 
 
-class _ApplyXFMInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                            fsl.preprocess.ApplyXFMInputSpec):
+class _ApplyXFMInputSpecRPT(
+    nrc._SVGReportCapableInputSpec, fsl.preprocess.ApplyXFMInputSpec
+):
     pass
 
 
@@ -198,16 +231,22 @@ else:
     _BBRegisterInputSpec = fs.preprocess.BBRegisterInputSpec6
 
 
-class _BBRegisterInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                              _BBRegisterInputSpec):
+class _BBRegisterInputSpecRPT(nrc._SVGReportCapableInputSpec, _BBRegisterInputSpec):
     # Adds default=True, usedefault=True
-    out_lta_file = traits.Either(traits.Bool, File, default=True, usedefault=True,
-                                 argstr="--lta %s", min_ver='5.2.0',
-                                 desc="write the transformation matrix in LTA format")
+    out_lta_file = traits.Either(
+        traits.Bool,
+        File,
+        default=True,
+        usedefault=True,
+        argstr="--lta %s",
+        min_ver="5.2.0",
+        desc="write the transformation matrix in LTA format",
+    )
 
 
-class _BBRegisterOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                               fs.preprocess.BBRegisterOutputSpec):
+class _BBRegisterOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, fs.preprocess.BBRegisterOutputSpec
+):
     pass
 
 
@@ -217,35 +256,39 @@ class BBRegisterRPT(nrc.RegistrationRC, fs.BBRegister):
 
     def _post_run_hook(self, runtime):
         outputs = self.aggregate_outputs(runtime=runtime)
-        mri_dir = os.path.join(self.inputs.subjects_dir,
-                               self.inputs.subject_id, 'mri')
-        target_file = os.path.join(mri_dir, 'brainmask.mgz')
+        mri_dir = os.path.join(self.inputs.subjects_dir, self.inputs.subject_id, "mri")
+        target_file = os.path.join(mri_dir, "brainmask.mgz")
 
         # Apply transform for simplicity
         mri_vol2vol = fs.ApplyVolTransform(
             source_file=self.inputs.source_file,
             target_file=target_file,
             lta_file=outputs.out_lta_file,
-            interp='nearest')
+            interp="nearest",
+        )
         res = mri_vol2vol.run()
 
         self._fixed_image = target_file
         self._moving_image = res.outputs.transformed_file
-        self._contour = os.path.join(mri_dir, 'ribbon.mgz')
+        self._contour = os.path.join(mri_dir, "ribbon.mgz")
         NIWORKFLOWS_LOG.info(
-            'Report - setting fixed (%s) and moving (%s) images',
-            self._fixed_image, self._moving_image)
+            "Report - setting fixed (%s) and moving (%s) images",
+            self._fixed_image,
+            self._moving_image,
+        )
 
         return super(BBRegisterRPT, self)._post_run_hook(runtime)
 
 
-class _MRICoregInputSpecRPT(nrc._SVGReportCapableInputSpec,
-                            fs.registration.MRICoregInputSpec):
+class _MRICoregInputSpecRPT(
+    nrc._SVGReportCapableInputSpec, fs.registration.MRICoregInputSpec
+):
     pass
 
 
-class _MRICoregOutputSpecRPT(reporting.ReportCapableOutputSpec,
-                             fs.registration.MRICoregOutputSpec):
+class _MRICoregOutputSpecRPT(
+    reporting.ReportCapableOutputSpec, fs.registration.MRICoregOutputSpec
+):
     pass
 
 
@@ -257,37 +300,41 @@ class MRICoregRPT(nrc.RegistrationRC, fs.MRICoreg):
         outputs = self.aggregate_outputs(runtime=runtime)
         mri_dir = None
         if isdefined(self.inputs.subject_id):
-            mri_dir = os.path.join(self.inputs.subjects_dir,
-                                   self.inputs.subject_id, 'mri')
+            mri_dir = os.path.join(
+                self.inputs.subjects_dir, self.inputs.subject_id, "mri"
+            )
 
         if isdefined(self.inputs.reference_file):
             target_file = self.inputs.reference_file
         else:
-            target_file = os.path.join(mri_dir, 'brainmask.mgz')
+            target_file = os.path.join(mri_dir, "brainmask.mgz")
 
         # Apply transform for simplicity
         mri_vol2vol = fs.ApplyVolTransform(
             source_file=self.inputs.source_file,
             target_file=target_file,
             lta_file=outputs.out_lta_file,
-            interp='nearest')
+            interp="nearest",
+        )
         res = mri_vol2vol.run()
 
         self._fixed_image = target_file
         self._moving_image = res.outputs.transformed_file
         if mri_dir is not None:
-            self._contour = os.path.join(mri_dir, 'ribbon.mgz')
+            self._contour = os.path.join(mri_dir, "ribbon.mgz")
         NIWORKFLOWS_LOG.info(
-            'Report - setting fixed (%s) and moving (%s) images',
-            self._fixed_image, self._moving_image)
+            "Report - setting fixed (%s) and moving (%s) images",
+            self._fixed_image,
+            self._moving_image,
+        )
 
         return super(MRICoregRPT, self)._post_run_hook(runtime)
 
 
 class _SimpleBeforeAfterInputSpecRPT(nrc._SVGReportCapableInputSpec):
-    before = File(exists=True, mandatory=True, desc='file before')
-    after = File(exists=True, mandatory=True, desc='file after')
-    wm_seg = File(desc='reference white matter segmentation mask')
+    before = File(exists=True, mandatory=True, desc="file before")
+    after = File(exists=True, mandatory=True, desc="file after")
+    wm_seg = File(desc="reference white matter segmentation mask")
     before_label = traits.Str("before", usedefault=True)
     after_label = traits.Str("after", usedefault=True)
 
@@ -303,14 +350,16 @@ class SimpleBeforeAfterRPT(nrc.RegistrationRC, nrc.ReportingInterface):
         self._moving_image = self.inputs.before
         self._contour = self.inputs.wm_seg if isdefined(self.inputs.wm_seg) else None
         NIWORKFLOWS_LOG.info(
-            'Report - setting before (%s) and after (%s) images',
-            self._fixed_image, self._moving_image)
+            "Report - setting before (%s) and after (%s) images",
+            self._fixed_image,
+            self._moving_image,
+        )
 
         return super(SimpleBeforeAfterRPT, self)._post_run_hook(runtime)
 
 
 class _ResampleBeforeAfterInputSpecRPT(_SimpleBeforeAfterInputSpecRPT):
-    base = traits.Enum('before', 'after', usedefault=True, mandatory=True)
+    base = traits.Enum("before", "after", usedefault=True, mandatory=True)
 
 
 class ResampleBeforeAfterRPT(SimpleBeforeAfterRPT):
@@ -319,111 +368,187 @@ class ResampleBeforeAfterRPT(SimpleBeforeAfterRPT):
     def _post_run_hook(self, runtime):
         self._fixed_image = self.inputs.after
         self._moving_image = self.inputs.before
-        if self.inputs.base == 'before':
+        if self.inputs.base == "before":
             resampled_after = nli.resample_to_img(self._fixed_image, self._moving_image)
-            fname = fname_presuffix(self._fixed_image, suffix='_resampled', newpath=runtime.cwd)
+            fname = fname_presuffix(
+                self._fixed_image, suffix="_resampled", newpath=runtime.cwd
+            )
             resampled_after.to_filename(fname)
             self._fixed_image = fname
         else:
-            resampled_before = nli.resample_to_img(self._moving_image, self._fixed_image)
-            fname = fname_presuffix(self._moving_image, suffix='_resampled', newpath=runtime.cwd)
+            resampled_before = nli.resample_to_img(
+                self._moving_image, self._fixed_image
+            )
+            fname = fname_presuffix(
+                self._moving_image, suffix="_resampled", newpath=runtime.cwd
+            )
             resampled_before.to_filename(fname)
             self._moving_image = fname
         self._contour = self.inputs.wm_seg if isdefined(self.inputs.wm_seg) else None
         NIWORKFLOWS_LOG.info(
-            'Report - setting before (%s) and after (%s) images',
-            self._fixed_image, self._moving_image)
+            "Report - setting before (%s) and after (%s) images",
+            self._fixed_image,
+            self._moving_image,
+        )
 
         runtime = super(ResampleBeforeAfterRPT, self)._post_run_hook(runtime)
-        NIWORKFLOWS_LOG.info('Successfully created report (%s)', self._out_report)
+        NIWORKFLOWS_LOG.info("Successfully created report (%s)", self._out_report)
         os.unlink(fname)
 
         return runtime
 
 
 class _EstimateReferenceImageInputSpec(BaseInterfaceInputSpec):
-    in_file = File(exists=True, mandatory=True, desc="4D EPI file")
-    sbref_file = File(exists=True, desc="Single band reference image")
+    in_file = InputMultiObject(
+        File(exists=True),
+        mandatory=True,
+        desc=(
+            "4D EPI file. If multiple files "
+            "are provided, they are assumed "
+            "to represent multiple echoes "
+            "from the same run."
+        ),
+    )
+    sbref_file = InputMultiObject(
+        File(exists=True),
+        desc=(
+            "Single band reference image. "
+            "If multiple files are provided, "
+            "they are assumed to represent "
+            "multiple echoes."
+        ),
+    )
     mc_method = traits.Enum(
-        "AFNI", "FSL", usedefault=True,
-        desc="Which software to use to perform motion correction")
+        "AFNI",
+        "FSL",
+        usedefault=True,
+        desc="Which software to use to perform motion correction",
+    )
+    multiecho = traits.Bool(
+        False,
+        usedefault=True,
+        desc=(
+            "If multiecho data was supplied, data from "
+            "the first echo will be selected."
+        ),
+    )
 
 
 class _EstimateReferenceImageOutputSpec(TraitedSpec):
     ref_image = File(exists=True, desc="3D reference image")
-    n_volumes_to_discard = traits.Int(desc="Number of detected non-steady "
-                                           "state volumes in the beginning of "
-                                           "the input file")
+    n_volumes_to_discard = traits.Int(
+        desc="Number of detected non-steady "
+        "state volumes in the beginning of "
+        "the input file"
+    )
 
 
 class EstimateReferenceImage(SimpleInterface):
     """
-    Given an 4D EPI file estimate an optimal reference image that could be later
-    used for motion estimation and coregistration purposes. If detected uses
-    T1 saturated volumes (non-steady state) otherwise a median of
+    Generate a reference 3D map from BOLD and SBRef EPI images for BOLD datasets.
+
+    Given a 4D BOLD file or one or more 3/4D SBRefs, estimate a reference
+    image for subsequent motion estimation and coregistration steps.
+    For the case of BOLD datasets, it estimates a number of T1w saturated volumes
+    (non-steady state at the beginning of the scan) and calculates the median
+    across them.
+    Otherwise (SBRefs or detected zero non-steady state frames), a median of
     of a subset of motion corrected volumes is used.
+    If the input reference (BOLD or SBRef) is 3D already, it just returns a
+    copy of the image with the NIfTI header extensions removed.
+
+    LIMITATION: If one wants to extract the reference from several SBRefs
+    with several echoes each, the first echo should be selected elsewhere
+    and run this interface in ``multiecho = False`` mode.
     """
+
     input_spec = _EstimateReferenceImageInputSpec
     output_spec = _EstimateReferenceImageOutputSpec
 
     def _run_interface(self, runtime):
-        ref_name = self.inputs.in_file
-        ref_nii = nb.load(ref_name)
-        n_volumes_to_discard = _get_vols_to_discard(ref_nii)
+        is_sbref = isdefined(self.inputs.sbref_file)
+        ref_input = self.inputs.sbref_file if is_sbref else self.inputs.in_file
 
-        self._results["n_volumes_to_discard"] = n_volumes_to_discard
-
-        out_ref_fname = os.path.join(runtime.cwd, "ref_bold.nii.gz")
-        if isdefined(self.inputs.sbref_file):
-            out_ref_fname = os.path.join(runtime.cwd, "ref_sbref.nii.gz")
-            ref_name = self.inputs.sbref_file
-            ref_nii = nb.squeeze_image(nb.load(ref_name))
-
-            # If reference is only 1 volume, return it directly
-            if len(ref_nii.shape) == 3:
-                ref_nii.header.extensions.clear()
-                ref_nii.to_filename(out_ref_fname)
-                self._results['ref_image'] = out_ref_fname
-                return runtime
+        if self.inputs.multiecho:
+            if len(ref_input) < 2:
+                input_name = "sbref_file" if is_sbref else "in_file"
+                raise ValueError("Argument 'multiecho' is True but "
+                                 f"'{input_name}' has only one element.")
             else:
-                # Reset this variable as it no longer applies
-                # and value for the output is stored in self._results
-                n_volumes_to_discard = 0
+                # Select only the first echo (see LIMITATION above for SBRefs)
+                ref_input = ref_input[:1]
+        elif not is_sbref and len(ref_input) > 1:
+            raise ValueError("Input 'in_file' cannot point to more than one file "
+                             "for single-echo BOLD datasets.")
+
+        # Build the nibabel spatial image we will work with
+        ref_im = []
+        for im_i in ref_input:
+            nib_i = nb.squeeze_image(nb.load(im_i))
+            if nib_i.dataobj.ndim == 3:
+                ref_im.append(nib_i)
+            elif nib_i.dataobj.ndim == 4:
+                ref_im += nb.four_to_three(nib_i)
+        ref_im = nb.squeeze_image(nb.concat_images(ref_im))
+
+        # Volumes to discard only makes sense with BOLD inputs.
+        if not is_sbref:
+            n_volumes_to_discard = _get_vols_to_discard(ref_im)
+            out_ref_fname = os.path.join(runtime.cwd, "ref_bold.nii.gz")
+        else:
+            n_volumes_to_discard = 0
+            out_ref_fname = os.path.join(runtime.cwd, "ref_sbref.nii.gz")
+
+        # Set interface outputs
+        self._results["n_volumes_to_discard"] = n_volumes_to_discard
+        self._results["ref_image"] = out_ref_fname
 
         # Slicing may induce inconsistencies with shape-dependent values in extensions.
         # For now, remove all. If this turns out to be a mistake, we can select extensions
         # that don't break pipeline stages.
-        ref_nii.header.extensions.clear()
+        ref_im.header.extensions.clear()
+
+        # If reference is only 1 volume, return it directly
+        if ref_im.dataobj.ndim == 3:
+            ref_im.to_filename(out_ref_fname)
+            return runtime
 
         if n_volumes_to_discard == 0:
-            if ref_nii.shape[-1] > 40:
-                ref_name = os.path.join(runtime.cwd, "slice.nii.gz")
-                nb.Nifti1Image(ref_nii.dataobj[:, :, :, 20:40], ref_nii.affine,
-                               ref_nii.header).to_filename(ref_name)
+            if ref_im.shape[-1] > 40:
+                ref_im = nb.Nifti1Image(
+                    ref_im.dataobj[:, :, :, 20:40], ref_im.affine, ref_im.header
+                )
 
+            ref_name = os.path.join(runtime.cwd, "slice.nii.gz")
+            ref_im.to_filename(ref_name)
             if self.inputs.mc_method == "AFNI":
-                res = afni.Volreg(in_file=ref_name, args='-Fourier -twopass',
-                                  zpad=4, outputtype='NIFTI_GZ').run()
+                res = afni.Volreg(
+                    in_file=ref_name,
+                    args="-Fourier -twopass",
+                    zpad=4,
+                    outputtype="NIFTI_GZ",
+                ).run()
             elif self.inputs.mc_method == "FSL":
-                res = fsl.MCFLIRT(in_file=ref_name,
-                                  ref_vol=0, interpolation='sinc').run()
+                res = fsl.MCFLIRT(
+                    in_file=ref_name, ref_vol=0, interpolation="sinc"
+                ).run()
             mc_slice_nii = nb.load(res.outputs.out_file)
 
             median_image_data = np.median(mc_slice_nii.get_fdata(), axis=3)
         else:
             median_image_data = np.median(
-                ref_nii.dataobj[:, :, :, :n_volumes_to_discard], axis=3)
+                ref_im.dataobj[:, :, :, :n_volumes_to_discard], axis=3
+            )
 
-        nb.Nifti1Image(median_image_data, ref_nii.affine,
-                       ref_nii.header).to_filename(out_ref_fname)
-
-        self._results["ref_image"] = out_ref_fname
-
+        nb.Nifti1Image(median_image_data, ref_im.affine, ref_im.header).to_filename(
+            out_ref_fname
+        )
         return runtime
 
 
 def _get_vols_to_discard(img):
     from nipype.algorithms.confounds import is_outlier
+
     data_slice = img.dataobj[:, :, :, :50]
     global_signal = data_slice.mean(axis=0).mean(axis=0).mean(axis=0)
     return is_outlier(global_signal)
