@@ -3,17 +3,21 @@
 """Interfaces under evaluation before upstreaming to nipype.interfaces.utility."""
 from nipype.interfaces.io import add_traits
 from nipype.interfaces.base import (
-    InputMultiObject, Str, DynamicTraitedSpec, BaseInterface, isdefined
+    InputMultiObject,
+    Str,
+    DynamicTraitedSpec,
+    BaseInterface,
+    isdefined,
 )
 
 
 class _KeySelectInputSpec(DynamicTraitedSpec):
-    key = Str(mandatory=True, desc='selective key')
-    keys = InputMultiObject(Str, mandatory=True, min=1, desc='index of keys')
+    key = Str(mandatory=True, desc="selective key")
+    keys = InputMultiObject(Str, mandatory=True, min=1, desc="index of keys")
 
 
 class _KeySelectOutputSpec(DynamicTraitedSpec):
-    key = Str(desc='propagates selected key')
+    key = Str(desc="propagates selected key")
 
 
 class KeySelect(BaseInterface):
@@ -85,23 +89,56 @@ class KeySelect(BaseInterface):
     <BLANKLINE>
 
     """
+
     input_spec = _KeySelectInputSpec
     output_spec = _KeySelectOutputSpec
 
     def __init__(self, keys=None, fields=None, **inputs):
+        """
+        Instantiate a KeySelect utility interface.
+
+        Examples
+        --------
+        >>> ks = KeySelect(fields='field1')
+        >>> ks.inputs
+        <BLANKLINE>
+        field1 = <undefined>
+        key = <undefined>
+        keys = <undefined>
+        <BLANKLINE>
+
+        >>> ks = KeySelect(fields='field1', field1=['a', 'b'])
+        >>> ks.inputs
+        <BLANKLINE>
+        field1 = ['a', 'b']
+        key = <undefined>
+        keys = <undefined>
+        <BLANKLINE>
+
+        >>> ks = KeySelect()
+        Traceback (most recent call last):
+        ValueError: A list or multiplexed...
+
+        >>> ks = KeySelect(fields='key')
+        Traceback (most recent call last):
+        ValueError: Some fields are invalid...
+
+        """
         # Call constructor
         super(KeySelect, self).__init__(**inputs)
 
         # Handle and initiate fields
         if not fields:
-            raise ValueError('A list or multiplexed fields must be provided at '
-                             'instantiation time.')
+            raise ValueError(
+                "A list or multiplexed fields must be provided at "
+                "instantiation time."
+            )
         if isinstance(fields, str):
             fields = [fields]
 
         _invalid = set(self.input_spec.class_editable_traits()).intersection(fields)
         if _invalid:
-            raise ValueError('Some fields are invalid (%s).' % ', '.join(_invalid))
+            raise ValueError("Some fields are invalid (%s)." % ", ".join(_invalid))
 
         self._fields = fields
 
@@ -120,7 +157,9 @@ class KeySelect(BaseInterface):
         if name == "keys":
             nitems = len(new)
             if len(set(new)) != nitems:
-                raise ValueError('Found duplicated entries in the index of ordered keys')
+                raise ValueError(
+                    "Found duplicated entries in the index of ordered keys"
+                )
 
         if not isdefined(self.inputs.keys):
             return
@@ -130,12 +169,15 @@ class KeySelect(BaseInterface):
 
         if name in self._fields:
             if isinstance(new, str) or len(new) < 1:
-                raise ValueError('Trying to set an invalid value (%s) for input "%s"' % (
-                                 new, name))
+                raise ValueError(
+                    'Trying to set an invalid value (%s) for input "%s"' % (new, name)
+                )
 
             if len(new) != len(self.inputs.keys):
-                raise ValueError('Length of value (%s) for input field "%s" does not match '
-                                 'the length of the indexing list.' % (new, name))
+                raise ValueError(
+                    'Length of value (%s) for input field "%s" does not match '
+                    "the length of the indexing list." % (new, name)
+                )
 
     def _run_interface(self, runtime):
         return runtime
@@ -143,14 +185,12 @@ class KeySelect(BaseInterface):
     def _list_outputs(self):
         index = self.inputs.keys.index(self.inputs.key)
 
-        outputs = {k: getattr(self.inputs, k)[index]
-                   for k in self._fields}
+        outputs = {k: getattr(self.inputs, k)[index] for k in self._fields}
 
-        outputs['key'] = self.inputs.key
+        outputs["key"] = self.inputs.key
         return outputs
 
     def _outputs(self):
         base = super(KeySelect, self)._outputs()
-        if self._fields:
-            base = add_traits(base, self._fields)
+        base = add_traits(base, self._fields)
         return base
