@@ -3,4 +3,4 @@
 
 from .core import generate_reports
 
-__all__ = ['generate_reports']
+__all__ = ["generate_reports"]
