@@ -8,30 +8,24 @@ Helpers for handling BIDS-like neuroimaging structures
 
 """
 from pathlib import Path
-import warnings
+import json
 import re
+import warnings
 from bids import BIDSLayout
-
-__all__ = ['BIDS_NAME']
-
-BIDS_NAME = re.compile(
-    r'^(.*\/)?'
-    '(?P<subject_id>sub-[a-zA-Z0-9]+)'
-    '(_(?P<session_id>ses-[a-zA-Z0-9]+))?'
-    '(_(?P<task_id>task-[a-zA-Z0-9]+))?'
-    '(_(?P<acq_id>acq-[a-zA-Z0-9]+))?'
-    '(_(?P<rec_id>rec-[a-zA-Z0-9]+))?'
-    '(_(?P<run_id>run-[a-zA-Z0-9]+))?')
+from packaging.version import Version
 
 
 class BIDSError(ValueError):
     def __init__(self, message, bids_root):
         indent = 10
         header = '{sep} BIDS root folder: "{bids_root}" {sep}'.format(
-            bids_root=bids_root, sep=''.join(['-'] * indent))
-        self.msg = '\n{header}\n{indent}{message}\n{footer}'.format(
-            header=header, indent=''.join([' '] * (indent + 1)),
-            message=message, footer=''.join(['-'] * len(header))
+            bids_root=bids_root, sep="".join(["-"] * indent)
+        )
+        self.msg = "\n{header}\n{indent}{message}\n{footer}".format(
+            header=header,
+            indent="".join([" "] * (indent + 1)),
+            message=message,
+            footer="".join(["-"] * len(header)),
         )
         super(BIDSError, self).__init__(self.msg)
         self.bids_root = bids_root
@@ -41,8 +35,9 @@ class BIDSWarning(RuntimeWarning):
     pass
 
 
-def collect_participants(bids_dir, participant_label=None, strict=False,
-                         bids_validate=True):
+def collect_participants(
+    bids_dir, participant_label=None, strict=False, bids_validate=True
+):
     """
     List the participants under the BIDS root and checks that participants
     designated with the participant_label argument exist in that folder.
@@ -83,11 +78,13 @@ def collect_participants(bids_dir, participant_label=None, strict=False,
     # Error: bids_dir does not contain subjects
     if not all_participants:
         raise BIDSError(
-            'Could not find participants. Please make sure the BIDS data '
-            'structure is present and correct. Datasets can be validated online '
-            'using the BIDS Validator (http://bids-standard.github.io/bids-validator/).\n'
-            'If you are using Docker for Mac or Docker for Windows, you '
-            'may need to adjust your "File sharing" preferences.', bids_dir)
+            "Could not find participants. Please make sure the BIDS data "
+            "structure is present and correct. Datasets can be validated online "
+            "using the BIDS Validator (http://bids-standard.github.io/bids-validator/).\n"
+            "If you are using Docker for Mac or Docker for Windows, you "
+            'may need to adjust your "File sharing" preferences.',
+            bids_dir,
+        )
 
     # No --participant-label was set, return all
     if not participant_label:
@@ -97,20 +94,26 @@ def collect_participants(bids_dir, participant_label=None, strict=False,
         participant_label = [participant_label]
 
     # Drop sub- prefixes
-    participant_label = [sub[4:] if sub.startswith('sub-') else sub for sub in participant_label]
+    participant_label = [
+        sub[4:] if sub.startswith("sub-") else sub for sub in participant_label
+    ]
     # Remove duplicates
     participant_label = sorted(set(participant_label))
     # Remove labels not found
     found_label = sorted(set(participant_label) & all_participants)
     if not found_label:
-        raise BIDSError('Could not find participants [{}]'.format(
-            ', '.join(participant_label)), bids_dir)
+        raise BIDSError(
+            "Could not find participants [{}]".format(", ".join(participant_label)),
+            bids_dir,
+        )
 
     # Warn if some IDs were not found
     notfound_label = sorted(set(participant_label) - all_participants)
     if notfound_label:
-        exc = BIDSError('Some participants were not found: {}'.format(
-            ', '.join(notfound_label)), bids_dir)
+        exc = BIDSError(
+            "Some participants were not found: {}".format(", ".join(notfound_label)),
+            bids_dir,
+        )
         if strict:
             raise exc
         warnings.warn(exc.msg, BIDSWarning)
@@ -118,8 +121,14 @@ def collect_participants(bids_dir, participant_label=None, strict=False,
     return found_label
 
 
-def collect_data(bids_dir, participant_label, task=None, echo=None,
-                 bids_validate=True):
+def collect_data(
+    bids_dir,
+    participant_label,
+    task=None,
+    echo=None,
+    bids_validate=True,
+    bids_filters=None,
+):
     """
     Uses pybids to retrieve the input data for a given participant
     >>> bids_root, _ = collect_data(str(datadir / 'ds054'), '100185',
@@ -146,6 +155,10 @@ def collect_data(bids_dir, participant_label, task=None, echo=None,
     ['.../ds054/sub-100185/anat/sub-100185_T1w.nii.gz']
     >>> bids_root['t2w']  # doctest: +ELLIPSIS
     []
+    >>> bids_root, _ = collect_data(str(datadir / 'ds051'), '01',
+    ...                             bids_validate=False, bids_filters={'t1w':{'run': 1}})
+    >>> bids_root['t1w']  # doctest: +ELLIPSIS
+    ['.../ds051/sub-01/anat/sub-01_run-01_T1w.nii.gz']
     """
     if isinstance(bids_dir, BIDSLayout):
         layout = bids_dir
@@ -153,29 +166,39 @@ def collect_data(bids_dir, participant_label, task=None, echo=None,
         layout = BIDSLayout(str(bids_dir), validate=bids_validate)
 
     queries = {
-        'fmap': {'datatype': 'fmap'},
-        'bold': {'datatype': 'func', 'suffix': 'bold'},
-        'sbref': {'datatype': 'func', 'suffix': 'sbref'},
-        'flair': {'datatype': 'anat', 'suffix': 'FLAIR'},
-        't2w': {'datatype': 'anat', 'suffix': 'T2w'},
-        't1w': {'datatype': 'anat', 'suffix': 'T1w'},
-        'roi': {'datatype': 'anat', 'suffix': 'roi'},
+        "fmap": {"datatype": "fmap"},
+        "bold": {"datatype": "func", "suffix": "bold"},
+        "sbref": {"datatype": "func", "suffix": "sbref"},
+        "flair": {"datatype": "anat", "suffix": "FLAIR"},
+        "t2w": {"datatype": "anat", "suffix": "T2w"},
+        "t1w": {"datatype": "anat", "suffix": "T1w"},
+        "roi": {"datatype": "anat", "suffix": "roi"},
     }
+    bids_filters = bids_filters or {}
+    for acq, entities in bids_filters.items():
+        queries[acq].update(entities)
 
     if task:
-        queries['bold']['task'] = task
+        queries["bold"]["task"] = task
 
     if echo:
-        queries['bold']['echo'] = echo
+        queries["bold"]["echo"] = echo
 
     subj_data = {
-        dtype: sorted(layout.get(return_type='file', subject=participant_label,
-                                 extension=['nii', 'nii.gz'], **query))
-        for dtype, query in queries.items()}
+        dtype: sorted(
+            layout.get(
+                return_type="file",
+                subject=participant_label,
+                extension=["nii", "nii.gz"],
+                **query,
+            )
+        )
+        for dtype, query in queries.items()
+    }
 
     # Special case: multi-echo BOLD, grouping echos
-    if any(['_echo-' in bold for bold in subj_data['bold']]):
-        subj_data['bold'] = group_multiecho(subj_data['bold'])
+    if any(["_echo-" in bold for bold in subj_data["bold"]]):
+        subj_data["bold"] = group_multiecho(subj_data["bold"])
 
     return subj_data, layout
 
@@ -192,8 +215,7 @@ def get_metadata_for_nifti(in_file, bids_dir=None, validate=True):
     >>>
 
     """
-    return _init_layout(in_file, bids_dir, validate).get_metadata(
-        str(in_file))
+    return _init_layout(in_file, bids_dir, validate).get_metadata(str(in_file))
 
 
 def _init_layout(in_file=None, bids_dir=None, validate=True):
@@ -203,12 +225,12 @@ def _init_layout(in_file=None, bids_dir=None, validate=True):
     if bids_dir is None:
         in_file = Path(in_file)
         for parent in in_file.parents:
-            if parent.name.startswith('sub-'):
+            if parent.name.startswith("sub-"):
                 bids_dir = parent.parent.resolve()
                 break
 
         if bids_dir is None:
-            raise RuntimeError('Could not infer BIDS root')
+            raise RuntimeError("Could not infer BIDS root")
 
     layout = BIDSLayout(str(bids_dir), validate=validate)
     return layout
@@ -292,7 +314,7 @@ def group_multiecho(bold_sess):
     from itertools import groupby
 
     def _grp_echos(x):
-        if '_echo-' not in x:
+        if "_echo-" not in x:
             return x
         echo = re.search("_echo-\\d*", x).group(0)
         return x.replace(echo, "_echo-?")
@@ -301,6 +323,89 @@ def group_multiecho(bold_sess):
     for _, bold in groupby(bold_sess, key=_grp_echos):
         bold = list(bold)
         # If single- or dual-echo, flatten list; keep list otherwise.
-        action = getattr(ses_uids, 'append' if len(bold) > 2 else 'extend')
+        action = getattr(ses_uids, "append" if len(bold) > 2 else "extend")
         action(bold)
     return ses_uids
+
+
+def relative_to_root(path):
+    """
+    Calculate the BIDS root folder given one file path's.
+
+    Example
+    -------
+    >>> str(relative_to_root(
+    ...     "/sub-03/sourcedata/sub-01/anat/sub-01_T1.nii.gz"
+    ... ))
+    'sub-01/anat/sub-01_T1.nii.gz'
+
+    >>> str(relative_to_root(
+    ...     "/sub-03/anat/sourcedata/sub-01/ses-preop/anat/sub-01_ses-preop_T1.nii.gz"
+    ... ))
+    'sub-01/ses-preop/anat/sub-01_ses-preop_T1.nii.gz'
+
+    >>> str(relative_to_root(
+    ...     "sub-01/anat/sub-01_T1.nii.gz"
+    ... ))
+    'sub-01/anat/sub-01_T1.nii.gz'
+
+    >>> str(relative_to_root("anat/sub-01_T1.nii.gz"))
+    'anat/sub-01_T1.nii.gz'
+
+    """
+    path = Path(path)
+    if path.name.startswith("sub-"):
+        parents = [path.name]
+        for p in path.parents:
+            parents.insert(0, p.name)
+            if p.name.startswith("sub-"):
+                return Path(*parents)
+        return path
+
+    raise ValueError(
+        f"Could not determine the BIDS root of <{path}>. "
+        "Only files under a subject directory are currently supported."
+    )
+
+
+def check_pipeline_version(cvers, data_desc):
+    """
+    Searches for existing BIDS pipeline output and compares against current pipeline version.
+
+    .. testsetup::
+
+    >>> import json
+    >>> data = {"PipelineDescription": {"Version": "1.1.1rc5"}}
+    >>> desc_file = Path(tmpdir) / 'sample_dataset_description.json'
+    >>> _ = desc_file.write_text(json.dumps(data))
+
+    Parameters
+    ----------
+    cvers : :obj:`str`
+        Current pipeline version
+    data_desc : :obj:`str` or :class:`os.Pathlike`
+        Path to pipeline output's dataset_description.json
+
+    Examples
+    --------
+    >>> check_pipeline_version('1.1.1rc5', 'sample_dataset_description.json') is None
+    True
+    >>> check_pipeline_version('1.1.1rc5+129.gbe0e5158', 'sample_dataset_description.json')
+
+    >>> check_pipeline_version('1.2', 'sample_dataset_description.json')  # doctest: +ELLIPSIS
+    'Previous output generated ...'
+
+    Returns
+    -------
+    message : :obj:`str` or :obj:`None`
+        A warning string if there is a difference between versions, otherwise ``None``.
+
+    """
+    data_desc = Path(data_desc)
+    if not data_desc.exists():
+        return
+
+    desc = json.loads(data_desc.read_text())
+    dvers = desc.get("PipelineDescription", {}).get("Version", "0+unknown")
+    if Version(cvers).public != Version(dvers).public:
+        return "Previous output generated by version {} found.".format(dvers)
