@@ -1,3 +1,4 @@
+"""Utilities to manipulate images."""
 import nibabel as nb
 import numpy as np
 
@@ -12,7 +13,7 @@ def unsafe_write_nifti_header_and_data(fname, header, data):
     with Fortran-ordered fire.
     """
     # ImageOpener handles zips transparently
-    with nb.openers.ImageOpener(fname, mode='wb') as fobj:
+    with nb.openers.ImageOpener(fname, mode="wb") as fobj:
         header.write_to(fobj)
         # This function serializes one block at a time to reduce memory usage a bit
         # It assumes Fortran-ordered data.
@@ -65,30 +66,35 @@ def overwrite_header(img, fname):
     header = img.header
     dataobj = img.dataobj
 
-    if getattr(img.dataobj, '_mmap', False):
+    if getattr(img.dataobj, "_mmap", False):
         raise ValueError("Image loaded with `mmap=True`. Aborting unsafe operation.")
 
     set_consumables(header, dataobj)
 
     ondisk = nb.load(fname, mmap=False)
 
-    try:
-        assert isinstance(ondisk.header, img.header_class)
-        # Check that the data block should be the same size
-        assert ondisk.get_data_dtype() == img.get_data_dtype()
-        assert img.header.get_data_shape() == ondisk.shape
-        # At the same offset from the start of the file
-        assert img.header['vox_offset'] == ondisk.dataobj.offset
-        # With the same scale factors
-        assert np.allclose(img.header['scl_slope'], ondisk.dataobj.slope, equal_nan=True)
-        assert np.allclose(img.header['scl_inter'], ondisk.dataobj.inter, equal_nan=True)
-    except AssertionError as e:
-        raise ValueError("Cannot write header without compromising data") from e
-    else:
-        data = np.asarray(dataobj.get_unscaled())
-        img._dataobj = data  # Allow old dataobj to be garbage collected
-        del ondisk, img, dataobj  # Drop everything we don't need, to be safe
-        unsafe_write_nifti_header_and_data(fname, header, data)
+    errmsg = "Cannot overwrite header (reason: {}).".format
+    if not isinstance(ondisk.header, img.header_class):
+        raise ValueError(errmsg("inconsistent header objects"))
+
+    if (
+        ondisk.get_data_dtype() != img.get_data_dtype()
+        or img.header.get_data_shape() != ondisk.shape
+    ):
+        raise ValueError(errmsg("data blocks are not the same size"))
+
+    if img.header["vox_offset"] != ondisk.dataobj.offset:
+        raise ValueError(errmsg("change in offset from start of file"))
+
+    if not np.allclose(
+        img.header["scl_slope"], ondisk.dataobj.slope, equal_nan=True
+    ) or not np.allclose(img.header["scl_inter"], ondisk.dataobj.inter, equal_nan=True):
+        raise ValueError(errmsg("change in scale factors"))
+
+    data = np.asarray(dataobj.get_unscaled())
+    img._dataobj = data  # Allow old dataobj to be garbage collected
+    del ondisk, img, dataobj  # Drop everything we don't need, to be safe
+    unsafe_write_nifti_header_and_data(fname, header, data)
 
 
 def update_header_fields(fname, **kwargs):
@@ -100,3 +106,104 @@ def update_header_fields(fname, **kwargs):
     for field, value in kwargs.items():
         img.header[field] = value
     overwrite_header(img, fname)
+
+
+def dseg_label(in_seg, label, newpath=None):
+    """Extract a particular label from a discrete segmentation."""
+    from pathlib import Path
+    import nibabel as nb
+    import numpy as np
+    from nipype.utils.filemanip import fname_presuffix
+
+    newpath = Path(newpath or ".")
+
+    nii = nb.load(in_seg)
+    data = np.int16(nii.dataobj) == label
+
+    out_file = fname_presuffix(in_seg, suffix="_mask", newpath=str(newpath.absolute()))
+    new = nii.__class__(data, nii.affine, nii.header)
+    new.set_data_dtype(np.uint8)
+    new.to_filename(out_file)
+    return out_file
+
+
+def resample_by_spacing(in_file, zooms, order=3, clip=True):
+    """Regrid the input image to match the new zooms."""
+    from pathlib import Path
+    import numpy as np
+    import nibabel as nb
+    from scipy.ndimage import map_coordinates
+
+    if isinstance(in_file, (str, Path)):
+        in_file = nb.load(in_file)
+
+    # Prepare output x-forms
+    sform, scode = in_file.get_sform(coded=True)
+    qform, qcode = in_file.get_qform(coded=True)
+
+    hdr = in_file.header.copy()
+    dtype = hdr.get_data_dtype()
+    data = np.asanyarray(in_file.dataobj)
+    zooms = np.array(zooms)
+
+    # Calculate the factors to normalize voxel size to the specific zooms
+    pre_zooms = np.array(in_file.header.get_zooms()[:3])
+
+    # Calculate an affine aligned with cardinal axes, for simplicity
+    card = nb.affines.from_matvec(np.diag(pre_zooms))
+    extent = card[:3, :3].dot(np.array(in_file.shape[:3]))
+    card[:3, 3] = -0.5 * extent
+
+    # Cover the FoV with the new grid
+    new_size = np.ceil(extent / zooms).astype(int)
+    offset = (extent - np.diag(zooms).dot(new_size)) * 0.5
+    new_card = nb.affines.from_matvec(np.diag(zooms), card[:3, 3] + offset)
+
+    # Calculate the new indexes
+    new_grid = np.array(
+        np.meshgrid(
+            np.arange(new_size[0]),
+            np.arange(new_size[1]),
+            np.arange(new_size[2]),
+            indexing="ij",
+        )
+    ).reshape((3, -1))
+
+    # Calculate the locations of the new samples, w.r.t. the original grid
+    ijk = np.linalg.inv(card).dot(
+        new_card.dot(np.vstack((new_grid, np.ones((1, new_grid.shape[1])))))
+    )
+
+    # Resample data in the new grid
+    resampled = map_coordinates(
+        data,
+        ijk[:3, :],
+        output=dtype,
+        order=order,
+        mode="constant",
+        cval=0,
+        prefilter=True,
+    ).reshape(new_size)
+    if clip:
+        resampled = np.clip(resampled, a_min=data.min(), a_max=data.max())
+
+    # Set new zooms
+    hdr.set_zooms(zooms)
+
+    # Get the original image's affine
+    affine = in_file.affine.copy()
+    # Determine rotations w.r.t. cardinal axis and eccentricity
+    rot = affine.dot(np.linalg.inv(card))
+    # Apply to the new cardinal, so that the resampling is consistent
+    new_affine = rot.dot(new_card)
+
+    if qcode != 0:
+        hdr.set_qform(new_affine.dot(np.linalg.inv(affine).dot(qform)), code=int(qcode))
+    if scode != 0:
+        hdr.set_sform(new_affine.dot(np.linalg.inv(affine).dot(sform)), code=int(scode))
+    if (scode, qcode) == (0, 0):
+        hdr.set_qform(new_affine, code=1)
+        hdr.set_sform(new_affine, code=1)
+
+    # Create a new x-form affine, aligned with cardinal axes, 1mm3 and centered.
+    return nb.Nifti1Image(resampled, new_affine, hdr)
