@@ -1,12 +1,19 @@
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
-"""
-Miscellaneous utilities
-"""
+"""Miscellaneous utilities."""
+import os
+import shutil
 
 
-__all__ = ['get_template_specs', 'fix_multi_T1w_source_name', 'add_suffix', 'read_crashfile',
-           'splitext', '_copy_any']
+__all__ = [
+    "get_template_specs",
+    "fix_multi_T1w_source_name",
+    "add_suffix",
+    "read_crashfile",
+    "splitext",
+    "_copy_any",
+    "clean_directory",
+]
 
 
 def get_template_specs(in_template, template_spec=None, default_resolution=1):
@@ -37,28 +44,37 @@ def get_template_specs(in_template, template_spec=None, default_resolution=1):
 
     """
     from templateflow.api import get as get_template
+
     # Massage spec (start creating if None)
     template_spec = template_spec or {}
-    template_spec['desc'] = template_spec.get('desc', None)
-    template_spec['atlas'] = template_spec.get('atlas', None)
-    template_spec['resolution'] = template_spec.pop(
-        'res', template_spec.get('resolution', default_resolution))
+    template_spec["desc"] = template_spec.get("desc", None)
+    template_spec["atlas"] = template_spec.get("atlas", None)
+    template_spec["resolution"] = template_spec.pop(
+        "res", template_spec.get("resolution", default_resolution)
+    )
 
-    common_spec = {'resolution': template_spec['resolution']}
-    if 'cohort' in template_spec:
-        common_spec['cohort'] = template_spec['cohort']
+    common_spec = {"resolution": template_spec["resolution"]}
+    if "cohort" in template_spec:
+        common_spec["cohort"] = template_spec["cohort"]
 
     tpl_target_path = get_template(in_template, **template_spec)
     if not tpl_target_path:
-        raise RuntimeError("""\
+        raise RuntimeError(
+            """\
 Could not find template "{0}" with specs={1}. Please revise your template \
-argument.""".format(in_template, template_spec))
+argument.""".format(
+                in_template, template_spec
+            )
+        )
 
     if isinstance(tpl_target_path, list):
-        raise RuntimeError("""\
+        raise RuntimeError(
+            """\
 The available template modifiers ({0}) did not select a unique template \
 (got "{1}"). Please revise your template argument.""".format(
-            template_spec, ', '.join([str(p) for p in tpl_target_path])))
+                template_spec, ", ".join([str(p) for p in tpl_target_path])
+            )
+        )
 
     return str(tpl_target_path), common_spec
 
@@ -75,6 +91,7 @@ def fix_multi_T1w_source_name(in_files):
     """
     import os
     from nipype.utils.filemanip import filename_to_list
+
     base, in_file = os.path.split(filename_to_list(in_files)[0])
     subject_label = in_file.split("_", 1)[0].split("-")[1]
     return os.path.join(base, "sub-%s_T1w.nii.gz" % subject_label)
@@ -92,30 +109,30 @@ def add_suffix(in_files, suffix):
     """
     import os.path as op
     from nipype.utils.filemanip import fname_presuffix, filename_to_list
-    return op.basename(fname_presuffix(filename_to_list(in_files)[0],
-                                       suffix=suffix))
+
+    return op.basename(fname_presuffix(filename_to_list(in_files)[0], suffix=suffix))
 
 
 def read_crashfile(path):
-    if path.endswith('.pklz'):
+    if path.endswith(".pklz"):
         return _read_pkl(path)
-    elif path.endswith('.txt'):
+    elif path.endswith(".txt"):
         return _read_txt(path)
-    raise RuntimeError('unknown crashfile format')
+    raise RuntimeError("unknown crashfile format")
 
 
 def _read_pkl(path):
     from nipype.utils.filemanip import loadcrash
+
     crash_data = loadcrash(path)
-    data = {'file': path,
-            'traceback': ''.join(crash_data['traceback'])}
-    if 'node' in crash_data:
-        data['node'] = crash_data['node']
-        if data['node'].base_dir:
-            data['node_dir'] = data['node'].output_dir()
+    data = {"file": path, "traceback": "".join(crash_data["traceback"])}
+    if "node" in crash_data:
+        data["node"] = crash_data["node"]
+        if data["node"].base_dir:
+            data["node_dir"] = data["node"].output_dir()
         else:
-            data['node_dir'] = "Node crashed before execution"
-        data['inputs'] = sorted(data['node'].inputs.trait_get().items())
+            data["node_dir"] = "Node crashed before execution"
+        data["inputs"] = sorted(data["node"].inputs.trait_get().items())
     return data
 
 
@@ -132,15 +149,16 @@ def _read_txt(path):
 
     """
     from pathlib import Path
+
     lines = Path(path).read_text().splitlines()
-    data = {'file': str(path)}
+    data = {"file": str(path)}
     traceback_start = 0
-    if lines[0].startswith('Node'):
-        data['node'] = lines[0].split(': ', 1)[1].strip()
-        data['node_dir'] = lines[1].split(': ', 1)[1].strip()
+    if lines[0].startswith("Node"):
+        data["node"] = lines[0].split(": ", 1)[1].strip()
+        data["node_dir"] = lines[1].split(": ", 1)[1].strip()
         inputs = []
-        cur_key = ''
-        cur_val = ''
+        cur_key = ""
+        cur_val = ""
         for i, line in enumerate(lines[5:]):
             if not line.strip():
                 continue
@@ -156,18 +174,21 @@ def _read_txt(path):
                 traceback_start = i + 5
                 break
 
-            cur_key, cur_val = tuple(line.split(' = ', 1))
+            cur_key, cur_val = tuple(line.split(" = ", 1))
 
-        data['inputs'] = sorted(inputs)
+        data["inputs"] = sorted(inputs)
     else:
-        data['node_dir'] = "Node crashed before execution"
-    data['traceback'] = '\n'.join(lines[traceback_start:]).strip()
+        data["node_dir"] = "Node crashed before execution"
+    data["traceback"] = "\n".join(lines[traceback_start:]).strip()
     return data
 
 
 def splitext(fname):
-    """Splits filename and extension (.gz safe)
+    """
+    Split filename in name and extension (.gz safe).
 
+    Examples
+    --------
     >>> splitext('some/file.nii.gz')
     ('file', '.nii.gz')
     >>> splitext('some/other/file.nii')
@@ -176,11 +197,18 @@ def splitext(fname):
     ('otherext', '.tar.gz')
     >>> splitext('text.txt')
     ('text', '.txt')
+    >>> splitext('some/figure.svg')
+    ('figure', '.svg')
+    >>> splitext('some/figure.svg.gz')
+    ('figure', '.svg.gz')
+    >>> splitext('some/sub-01_bold.func.gii')
+    ('sub-01_bold.func', '.gii')
 
     """
     from pathlib import Path
+
     basename = str(Path(fname).name)
-    stem = Path(basename.rstrip('.gz')).stem
+    stem = Path(basename.rstrip(".gz")).stem
     return stem, basename[len(stem):]
 
 
@@ -190,8 +218,8 @@ def _copy_any(src, dst):
     from shutil import copyfileobj
     from nipype.utils.filemanip import copyfile
 
-    src_isgz = src.endswith('.gz')
-    dst_isgz = dst.endswith('.gz')
+    src_isgz = src.endswith(".gz")
+    dst_isgz = dst.endswith(".gz")
     if not src_isgz and not dst_isgz:
         copyfile(src, dst, copy=True, use_hardlink=True)
         return False  # Make sure we do not reuse the hardlink later
@@ -201,11 +229,11 @@ def _copy_any(src, dst):
         os.unlink(dst)
 
     src_open = gzip.open if src_isgz else open
-    with src_open(src, 'rb') as f_in:
-        with open(dst, 'wb') as f_out:
+    with src_open(src, "rb") as f_in:
+        with open(dst, "wb") as f_out:
             if dst_isgz:
-                # Remove FNAME header from gzip (poldracklab/fmriprep#1480)
-                gz_out = gzip.GzipFile('', 'wb', 9, f_out, 0.)
+                # Remove FNAME header from gzip (nipreps/fmriprep#1480)
+                gz_out = gzip.GzipFile("", "wb", 9, f_out, 0.0)
                 copyfileobj(f_in, gz_out)
                 gz_out.close()
             else:
@@ -214,5 +242,78 @@ def _copy_any(src, dst):
     return True
 
 
-if __name__ == '__main__':
+def clean_directory(path):
+    """
+    Clears a directory of all contents.
+
+    Returns `True` if no content remains. If any content cannot be removed, returns `False`.
+
+    Notes
+    -----
+    This function is not guaranteed to work across multiple threads or processes.
+
+    """
+    for f in os.scandir(path):  # switch to context manager when py35 dropped
+        if f.is_file() or f.is_symlink():
+            try:
+                os.unlink(f.path)
+            except OSError:
+                return False
+        elif f.is_dir():
+            try:
+                shutil.rmtree(f.path)
+            except OSError:
+                return False
+    return True
+
+
+def pass_dummy_scans(algo_dummy_scans, dummy_scans=None):
+    """
+    Graft manually provided number of dummy scans, if necessary.
+
+    Parameters
+    ----------
+    algo_dummy_scans : int
+        number of volumes to skip determined by an algorithm
+    dummy_scans : int or None
+        number of volumes to skip determined by the user
+
+    Returns
+    -------
+    skip_vols_num : int
+        number of volumes to skip
+
+    """
+    if dummy_scans is None:
+        return algo_dummy_scans
+    return dummy_scans
+
+
+def check_valid_fs_license():
+    """
+    Run ``mri_convert`` to assess FreeSurfer access to a license.
+
+    Returns
+    -------
+    valid : :obj:`bool`
+        FreeSurfer successfully executed (valid license)
+
+    """
+    from pathlib import Path
+    import subprocess as sp
+    from tempfile import TemporaryDirectory
+    from pkg_resources import resource_filename
+
+    with TemporaryDirectory() as tmpdir:
+        # quick FreeSurfer command
+        _cmd = (
+            "mri_convert",
+            resource_filename("niworkflows", "data/sentinel.nii.gz"),
+            str(Path(tmpdir) / "out.mgz"),
+        )
+        proc = sp.run(_cmd, stdout=sp.PIPE, stderr=sp.STDOUT)
+    return proc.returncode == 0 and "ERROR:" not in proc.stdout.decode()
+
+
+if __name__ == "__main__":
     pass
