@@ -1,9 +1,6 @@
-# -*- coding: utf-8 -*-
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
-from __future__ import absolute_import, division, print_function, unicode_literals
-
 from .plots import plot_carpet
 from .utils import SVGNS
 
-__all__ = ['plot_carpet', 'SVGNS']
+__all__ = ["plot_carpet", "SVGNS"]
